% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-airmo.R
\name{airmo_wide_to_rmo_long}
\alias{airmo_wide_to_rmo_long}
\title{Converts the wide format from the files to rmo long}
\usage{
airmo_wide_to_rmo_long(header, data, tz = "Etc/GMT-1", time_shift = NULL)
}
\arguments{
\item{header}{dataframe containing airmo_kurzname, parameter, zeitfenster and einheit}

\item{data}{dataframe only containing the data from the file}

\item{tz}{time zone of date field. Be carefull Etc/GMT + == -. Default "Etc/GMT-1"}

\item{time_shift}{a lubridate period to add to the time. Default NULL}
}
\value{
dataframe in rmo long format
}
\description{
Converts the wide format from the files to rmo long
}
\seealso{
\itemize{
\item \code{\link[lubridate]{period}} - Create or parse period objects
\item \code{\link[base]{timezones}} - Information about time zones in R
}
}
\keyword{internal}
