% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{class}
\name{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction}
\title{Class for the \emph{\eqn{\alpha}-stable Bernstein function}}
\description{
For the \eqn{\alpha}-stable Lévy subordinator with \eqn{0 < \alpha < 1},
the corresponding Bernstein function is the power function with exponent
\eqn{\alpha}, i.e.
\deqn{
  \psi(x) = x^\alpha, x > 0.
}
}
\details{
For the \eqn{\alpha}-stable Bernstein function, the higher order
alternatingiterated forward differences are not known in closed form. But we
can use numerical integration (here: \code{\link[stats]{integrate}}) to
approximate it with the following representation:
\deqn{
  {(-1)}^{k} \Delta^k \psi(x) = \int_0^\infty e^{-ux} (1-e^{-u})^k \alpha / \Gamma(1-\alpha) u^{-1-\alpha} du, x > 0 .
}
}
\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}.}
}}

\examples{
bf <- AlphaStableBernsteinFunction(alpha=0.5)

}
\seealso{
\linkS4class{BernsteinFunction}
}
