% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{AlphaStableBernsteinFunction-constructor}
\alias{AlphaStableBernsteinFunction-constructor}
\alias{initialize,AlphaStableBernsteinFunction-method}
\alias{initialize,AlphaStableBernsteinFunction,ANY-method}
\title{Constructor of \emph{\eqn{\alpha}-stable Bernstein function}}
\usage{
\S4method{initialize}{AlphaStableBernsteinFunction}(.Object, alpha = log2(2 - 0.5))
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}
}
\description{
Constructor of \emph{\eqn{\alpha}-stable Bernstein function}
}
\examples{
AlphaStableBernsteinFunction()
AlphaStableBernsteinFunction(alpha = 0.5)

}
\seealso{
\linkS4class{AlphaStableBernsteinFunction}
}
