% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-calcExShockArrivalIntensities.R
\name{calcExShockArrivalIntensities}
\alias{calcExShockArrivalIntensities}
\title{Calculate exchangeable shock-arrival intensities}
\usage{
calcExShockArrivalIntensities(object, d, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}

\item{d}{A positive integer, larger than two, for the \emph{dimension}.}

\item{cscale}{A positive number for the \emph{composite scaling factor}.}

\item{...}{pass-through parameter.}
}
\description{
Calculates (unscaled) \emph{exchangeable shock-arrival intensities}, see \code{\link[=rmo]{rmo()}}.
}
\details{
For a given Bernstein function, the exchangeable shock-arrival intensities
are defined as
\deqn{
  \lambda_{i}
    = {(-1)}^{i-1} \Delta^{i}{ \psi{(d-i)} } ,
      \quad 1 \leq i \leq d .
}
The calculation of the exchangeable shock-arrival intensities using this
formula is usually not numerically stable. Consequently, the various
alternative approaches are used dependent on the class of the Bernstein
function.
}
\examples{
bf <- AlphaStableBernsteinFunction(alpha = 0.4)
calcExShockArrivalIntensities(bf, 3)
}
\seealso{
\code{\link[=rmo]{rmo()}}

Other Bernstein function generics: 
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{getDefaultMethodString}()},
\code{\link{getLevyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{valueOf}()},
\code{\link{valueOf0}()}
}
\concept{Bernstein function generics}
