% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-calcValue.R
\name{calcValue}
\alias{calcValue}
\title{Evaluate Bernstein functions}
\usage{
calcValue(object, x, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}

\item{x}{A nonnegative numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{cscale}{A positive number for the \emph{composite scaling factor}.}

\item{...}{Pass-through parameter.}
}
\description{
Calculate the values for a Bernstein function function.
}
\examples{
bf <- AlphaStableBernsteinFunction(alpha = 0.7)
calcValue(bf, 1:5)
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcIterativeDifference}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{getDefaultMethodString}()},
\code{\link{getLevyDensity}()},
\code{\link{getStieltjesDensity}()}
}
\concept{Bernstein function generics}
