% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-CompleteBernsteinFunction.R
\docType{class}
\name{CompleteBernsteinFunction-class}
\alias{CompleteBernsteinFunction-class}
\title{Virtual superclass for complete Bernstein functions}
\description{
A virtual superclass for all Bernstein functions which can represented
by a Stieltjes density (no drift or killing rate). That means that there
exists a Stieltjes measure \eqn{\sigma} such that
\deqn{
  \psi(x) = \int_0^\infty \frac{x}{x + u} \sigma(du) , x > 0 .
}
}
\seealso{
\linkS4class{LevyBernsteinFunction},
\linkS4class{BernsteinFunction}
\code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{BernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Stieltjes Bernstein function classes}
\concept{Virtual Bernstein function classes}
