% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ConstantBernsteinFunction.R
\docType{class}
\name{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction}
\alias{initialize,ConstantBernsteinFunction-method}
\alias{initialize,ConstantBernsteinFunction,ANY-method}
\alias{show,ConstantBernsteinFunction-method}
\alias{valueOf,ConstantBernsteinFunction-method}
\alias{valueOf,ConstantBernsteinFunction,ANY-method}
\title{Class for (almost) constant Bernstein functions}
\usage{
\S4method{initialize}{ConstantBernsteinFunction}(.Object, constant)

\S4method{show}{ConstantBernsteinFunction}(object)

\S4method{valueOf}{ConstantBernsteinFunction}(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{constant}{Non-negative number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{a non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{a non-negative integer with the order of the alternating iterated
forward differences taken on the Bernstein function.}

\item{n, k}{non-negative numbers for the binomial factor.}

\item{cscale}{a positive numeric scalar with the composite scaling factor.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
A \emph{constant Bernstein function} is a Bernstein function with only a
constant part (for \eqn{x > 0}), i.e. \eqn{b = 0} and \eqn{\nu = 0}. In
particular,
\deqn{
  \psi(x) = a , x > 0
}
}
\section{Functions}{
\itemize{
\item \code{initialize(ConstantBernsteinFunction)}: Constructor

\item \code{show(ConstantBernsteinFunction)}: Display the object.

\item \code{valueOf(ConstantBernsteinFunction)}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}

}}
\section{Slots}{

\describe{
\item{\code{constant}}{The non-negative \emph{killing} parameter (i.e. \eqn{a}
in the representation)}
}}

\examples{
ConstantBernsteinFunction()
ConstantBernsteinFunction(constant = 0.2)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
