% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{ConstantBernsteinFunction-constructor}
\alias{ConstantBernsteinFunction-constructor}
\alias{initialize,ConstantBernsteinFunction-method}
\alias{initialize,ConstantBernsteinFunction,ANY-method}
\title{Constructor for `constant Bernstein functions*}
\usage{
\S4method{initialize}{ConstantBernsteinFunction}(.Object, constant = 1)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{constant}{Non-negative number.}
}
\description{
Constructor for `constant Bernstein functions*
}
\examples{
ConstantBernsteinFunction()
ConstantBernsteinFunction(constant = 0.2)

}
\seealso{
\linkS4class{ConstantBernsteinFunction}
}
