% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ConvexCombinationOfBernsteinFunctions.R
\docType{class}
\name{ConvexCombinationOfBernsteinFunctions-class}
\alias{ConvexCombinationOfBernsteinFunctions-class}
\alias{ConvexCombinationOfBernsteinFunctions}
\alias{initialize,ConvexCombinationOfBernsteinFunctions-method}
\alias{initialize,ConvexCombinationOfBernsteinFunctions,ANY-method}
\alias{show,ConvexCombinationOfBernsteinFunctions-method}
\alias{valueOf,ConvexCombinationOfBernsteinFunctions-method}
\alias{valueOf,ConvexCombinationOfBernsteinFunctions,ANY-method}
\title{Class for convex combinations of Bernstein functions}
\usage{
\S4method{initialize}{ConvexCombinationOfBernsteinFunctions}(.Object, coefficients, points)

\S4method{show}{ConvexCombinationOfBernsteinFunctions}(object)

\S4method{valueOf}{ConvexCombinationOfBernsteinFunctions}(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{coefficients}{Derives from \linkS4class{ConvexCombinationOfBernsteinFunctions}.}

\item{points}{Derives from \linkS4class{ConvexCombinationOfBernsteinFunctions}.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{a non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{a non-negative integer with the order of the alternating iterated
forward differences taken on the Bernstein function.}

\item{n, k}{non-negative numbers for the binomial factor.}

\item{cscale}{a positive numeric scalar with the composite scaling factor.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
Bernstein functions are stable under convex combinations, i.e. if
\eqn{\psi_1, \ldots, \psi_n} are Bernstein functions and
\eqn{c_1, \ldots, c_n > 0} are positive real values,
\deqn{
  x \mapsto c_1 \psi_1(x) + \cdots + c_n \psi_n(x) , x>0,
}
is also a Bernstein function.
}
\section{Functions}{
\itemize{
\item \code{initialize(ConvexCombinationOfBernsteinFunctions)}: Constructor

\item \code{show(ConvexCombinationOfBernsteinFunctions)}: Display the object.

\item \code{valueOf(ConvexCombinationOfBernsteinFunctions)}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}

}}
\section{Slots}{

\describe{
\item{\code{coefficients}}{Numeric vector of positive real values.}

\item{\code{points}}{List of Bernstein functions.}
}}

\examples{
ConvexCombinationOfBernsteinFunctions()
bf1 <- LinearBernsteinFunction(scale = 0.2)
bf2 <- ConstantBernsteinFunction(constant = 0.5)
bf3 <- AlphaStableBernsteinFunction(alpha = 0.5)
ConvexCombinationOfBernsteinFunctions(
    coefficients = c(0.2, 0.5, 0.1),
    points = list(bf1, bf2, bf3)
)
}
