% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-defaultMethod.R
\name{defaultMethod}
\alias{defaultMethod}
\title{Default method for approximating Bernstein function differences}
\usage{
defaultMethod(object)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}
}
\value{
A String indicating the default method used for the approximation.
}
\description{
Evaluating Bernstein function differences is usually not numerically stable.
Consequently, various alternative approaches are used dependent on the class
of the Bernstein function. This method returns a String indicating the
default method used for the approximation.
}
\examples{
defaultMethod(AlphaStableBernsteinFunction(alpha = 0.7))
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{levyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{valueOf}()},
\code{\link{valueOf0}()}
}
\concept{Bernstein function generics}
