% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{ExponentialBernsteinFunction-constructor}
\alias{ExponentialBernsteinFunction-constructor}
\alias{initialize,ExponentialBernsteinFunction-method}
\alias{initialize,ExponentialBernsteinFunction,ANY-method}
\title{Constructor of the \emph{Exponential Bernstein function}}
\usage{
\S4method{initialize}{ExponentialBernsteinFunction}(.Object, lambda = 1)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{lambda}{Positive number.}
}
\description{
Constructor of the \emph{Exponential Bernstein function}
}
\examples{
ExponentialBernsteinFunction()
ExponentialBernsteinFunction(lambda = 0.5)

}
\seealso{
\linkS4class{ExponentialBernsteinFunction}
}
