% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-GammaBernsteinFunction.R
\docType{class}
\name{GammaBernsteinFunction-class}
\alias{GammaBernsteinFunction-class}
\alias{GammaBernsteinFunction}
\title{Class for Gamma Bernstein functions}
\description{
The \emph{Gamma Bernstein function}, is the Bernstein function of a
subordinator with a (scaled) Gamma distribution. The representation is for
\eqn{a > 0}
\deqn{
  \psi(x) = \log(1 + \frac{x}{a}), x > 0.
}
}
\details{
For this Bernstein function, the higher-order alternating iterated forward
differences are known in closed form but cannot be evaluated numerically
without the danger of loss of significance. But we can use numerical
integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^{k} \psi(x)
    = \int_{0}^{\infty} e^{-ux} {(1 - e^{-u})}^{k}
      \frac{e^{-au}}{u} du, x>0, k>0.
}

This Bernstein function is no. 26 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
\subsection{Lévy density}{

\deqn{
  \nu(du)
    = \frac{\operatorname{e}^{-a u}}{u}, \quad u > 0 .
}
}

\subsection{Stieltjes density}{

\deqn{
  \sigma(du)
    = 1 / u du, u > a .
}
}
}
\section{Slots}{

\describe{
\item{\code{a}}{Scale parameter for the Lévy measure.}
}}

\examples{
GammaBernsteinFunction()
GammaBernsteinFunction(a = 2)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction}, \code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}

Other Complete Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Complete Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Logarithmic Bernstein function classes}
