% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{GammaBernsteinFunction-constructor}
\alias{GammaBernsteinFunction-constructor}
\alias{initialize,GammaBernsteinFunction-method}
\alias{initialize,GammaBernsteinFunction,ANY-method}
\title{Constructor of the \emph{\eqn{\Gamma} Bernstein function}}
\usage{
\S4method{initialize}{GammaBernsteinFunction}(.Object, a = 1)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{a}{Positive number}
}
\description{
Constructor of the \emph{\eqn{\Gamma} Bernstein function}
}
\examples{
GammaBernsteinFunction()
GammaBernsteinFunction(a = 2)

}
\seealso{
\linkS4class{GammaBernsteinFunction}
}
