% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-getDefaultMethodString.R
\name{getDefaultMethodString}
\alias{getDefaultMethodString}
\title{Default method for approximating Bernstein function differences}
\usage{
getDefaultMethodString(object)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}
}
\value{
A String indicating the default method used for the approximation.
}
\description{
Evaluating Bernstein function differences is usually not numerically stable.
Consequently, various alternative approaches are used dependent on the class
of the Bernstein function. This method returns a String indicating the
default method used for the approximation.
}
\examples{
getDefaultMethodString(AlphaStableBernsteinFunction(alpha = 0.7))
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcIterativeDifference}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{calcValue}()},
\code{\link{getLevyDensity}()},
\code{\link{getStieltjesDensity}()}
}
\concept{Bernstein function generics}
