% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-getLevyDensity.R
\name{getLevyDensity}
\alias{getLevyDensity}
\title{Return the Lévy density of a Bernstein function}
\usage{
getLevyDensity(object)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}.}
}
\description{
Returns the \emph{Lévy density} with \code{lower}, \code{upper}, and \code{type} attributes if
continuous and returns a \code{data.frame} with named columns \code{x} (atoms) and \code{y}
(weights) as well as a type attribute if discrete. The \code{type} attribute is
either \code{"continuous"} or \code{"discrete"}.
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{getDefaultMethodString}()},
\code{\link{getStieltjesDensity}()},
\code{\link{valueOf}()},
\code{\link{valueOf0}()}
}
\concept{Bernstein function generics}
