% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-getStieltjesDensity.R
\name{getStieltjesDensity}
\alias{getStieltjesDensity}
\title{Return the Stieltjes density of a Bernstein function}
\usage{
getStieltjesDensity(object)
}
\arguments{
\item{object}{An object deriving from \linkS4class{CompleteBernsteinFunction}.}
}
\description{
Returns the \emph{Stieltjes density} with \code{lower}, \code{upper}, and \code{type} attributes
if continuous and returns a \code{data.frame} with named columns \code{x} (atoms) and
\code{y} (weights) as well as a type attribute if discrete. The \code{type} attribute
is either \code{"continuous"} or \code{"discrete"}.
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcIterativeDifference}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{calcValue}()},
\code{\link{getDefaultMethodString}()},
\code{\link{getLevyDensity}()}
}
\concept{Bernstein function generics}
