% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BernsteinFunction.R,
%   R/s4-LevyBernsteinFunction.R, R/s4-CompleteBernsteinFunction.R,
%   R/s4-AlphaStableBernsteinFunction.R,
%   R/s4-CompositeScaledBernsteinFunction.R, R/s4-ConstantBernsteinFunction.R,
%   R/s4-ConvexCombinationOfBernsteinFunctions.R,
%   R/s4-ExponentialBernsteinFunction.R, R/s4-GammaBernsteinFunction.R,
%   R/s4-InverseGaussianBernsteinFunction.R, R/s4-LinearBernsteinFunction.R,
%   R/s4-ParetoBernsteinFunction.R, R/s4-PoissonBernsteinFunction.R,
%   R/s4-ScaledBernsteinFunction.R, R/s4-SumOfBernsteinFunctions.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{calcExShockArrivalIntensities,BernsteinFunction-method}
\alias{calcExShockSizeArrivalIntensities,BernsteinFunction-method}
\alias{calcShockArrivalIntensities,BernsteinFunction-method}
\alias{calcMDCMGeneratorMatrix,BernsteinFunction-method}
\alias{getDefaultMethodString,LevyBernsteinFunction-method}
\alias{calcIterativeDifference,LevyBernsteinFunction-method}
\alias{calcValue,LevyBernsteinFunction-method}
\alias{calcIterativeDifference,CompleteBernsteinFunction-method}
\alias{initialize,AlphaStableBernsteinFunction-method}
\alias{show,AlphaStableBernsteinFunction-method}
\alias{getLevyDensity,AlphaStableBernsteinFunction-method}
\alias{getStieltjesDensity,AlphaStableBernsteinFunction-method}
\alias{calcValue,AlphaStableBernsteinFunction-method}
\alias{initialize,CompositeScaledBernsteinFunction-method}
\alias{show,CompositeScaledBernsteinFunction-method}
\alias{calcValue,CompositeScaledBernsteinFunction-method}
\alias{calcIterativeDifference,CompositeScaledBernsteinFunction-method}
\alias{initialize,ConstantBernsteinFunction-method}
\alias{show,ConstantBernsteinFunction-method}
\alias{calcValue,ConstantBernsteinFunction-method}
\alias{calcIterativeDifference,ConstantBernsteinFunction-method}
\alias{initialize,ConvexCombinationOfBernsteinFunctions-method}
\alias{show,ConvexCombinationOfBernsteinFunctions-method}
\alias{calcValue,ConvexCombinationOfBernsteinFunctions-method}
\alias{calcIterativeDifference,ConvexCombinationOfBernsteinFunctions-method}
\alias{initialize,ExponentialBernsteinFunction-method}
\alias{show,ExponentialBernsteinFunction-method}
\alias{getLevyDensity,ExponentialBernsteinFunction-method}
\alias{getStieltjesDensity,ExponentialBernsteinFunction-method}
\alias{getDefaultMethodString,ExponentialBernsteinFunction-method}
\alias{calcValue,ExponentialBernsteinFunction-method}
\alias{initialize,GammaBernsteinFunction-method}
\alias{show,GammaBernsteinFunction-method}
\alias{getLevyDensity,GammaBernsteinFunction-method}
\alias{getStieltjesDensity,GammaBernsteinFunction-method}
\alias{calcValue,GammaBernsteinFunction-method}
\alias{initialize,InverseGaussianBernsteinFunction-method}
\alias{show,InverseGaussianBernsteinFunction-method}
\alias{getLevyDensity,InverseGaussianBernsteinFunction-method}
\alias{getStieltjesDensity,InverseGaussianBernsteinFunction-method}
\alias{calcValue,InverseGaussianBernsteinFunction-method}
\alias{initialize,LinearBernsteinFunction-method}
\alias{show,LinearBernsteinFunction-method}
\alias{calcValue,LinearBernsteinFunction-method}
\alias{calcIterativeDifference,LinearBernsteinFunction-method}
\alias{initialize,ParetoBernsteinFunction-method}
\alias{show,ParetoBernsteinFunction-method}
\alias{getLevyDensity,ParetoBernsteinFunction-method}
\alias{calcValue,ParetoBernsteinFunction-method}
\alias{initialize,PoissonBernsteinFunction-method}
\alias{show,PoissonBernsteinFunction-method}
\alias{getLevyDensity,PoissonBernsteinFunction-method}
\alias{calcValue,PoissonBernsteinFunction-method}
\alias{initialize,ScaledBernsteinFunction-method}
\alias{show,ScaledBernsteinFunction-method}
\alias{calcValue,ScaledBernsteinFunction-method}
\alias{calcIterativeDifference,ScaledBernsteinFunction-method}
\alias{initialize,SumOfBernsteinFunctions-method}
\alias{show,SumOfBernsteinFunctions-method}
\alias{calcValue,SumOfBernsteinFunctions-method}
\alias{calcIterativeDifference,SumOfBernsteinFunctions-method}
\title{For S4 methods that require a documentation entry but only clutter the index.}
\usage{
\S4method{calcExShockArrivalIntensities}{BernsteinFunction}(object, d, cscale = 1, ...)

\S4method{calcExShockSizeArrivalIntensities}{BernsteinFunction}(object, d, cscale = 1, ...)

\S4method{calcShockArrivalIntensities}{BernsteinFunction}(object, d, cscale = 1, ...)

\S4method{calcMDCMGeneratorMatrix}{BernsteinFunction}(object, d, cscale = 1, ...)

\S4method{getDefaultMethodString}{LevyBernsteinFunction}(object)

\S4method{calcIterativeDifference}{LevyBernsteinFunction}(
  object,
  x,
  difference_order,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...,
  method = c("default", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{calcValue}{LevyBernsteinFunction}(object, x, cscale = 1, method = getDefaultMethodString(object), ...)

\S4method{calcIterativeDifference}{CompleteBernsteinFunction}(
  object,
  x,
  difference_order,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{initialize}{AlphaStableBernsteinFunction}(.Object, alpha)

\S4method{show}{AlphaStableBernsteinFunction}(object)

\S4method{getLevyDensity}{AlphaStableBernsteinFunction}(object)

\S4method{getStieltjesDensity}{AlphaStableBernsteinFunction}(object)

\S4method{calcValue}{AlphaStableBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{CompositeScaledBernsteinFunction}(.Object, cscale, original)

\S4method{show}{CompositeScaledBernsteinFunction}(object)

\S4method{calcValue}{CompositeScaledBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{CompositeScaledBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  n = 1,
  k = 0,
  cscale = 1,
  ...
)

\S4method{initialize}{ConstantBernsteinFunction}(.Object, constant)

\S4method{show}{ConstantBernsteinFunction}(object)

\S4method{calcValue}{ConstantBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{ConstantBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...
)

\S4method{initialize}{ConvexCombinationOfBernsteinFunctions}(.Object, coefficients, points)

\S4method{show}{ConvexCombinationOfBernsteinFunctions}(object)

\S4method{calcValue}{ConvexCombinationOfBernsteinFunctions}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{ConvexCombinationOfBernsteinFunctions}(
  object,
  x,
  difference_order = 0L,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...
)

\S4method{initialize}{ExponentialBernsteinFunction}(.Object, lambda)

\S4method{show}{ExponentialBernsteinFunction}(object)

\S4method{getLevyDensity}{ExponentialBernsteinFunction}(object)

\S4method{getStieltjesDensity}{ExponentialBernsteinFunction}(object)

\S4method{getDefaultMethodString}{ExponentialBernsteinFunction}(object)

\S4method{calcValue}{ExponentialBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{GammaBernsteinFunction}(.Object, a)

\S4method{show}{GammaBernsteinFunction}(object)

\S4method{getLevyDensity}{GammaBernsteinFunction}(object)

\S4method{getStieltjesDensity}{GammaBernsteinFunction}(object)

\S4method{calcValue}{GammaBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{InverseGaussianBernsteinFunction}(.Object, eta)

\S4method{show}{InverseGaussianBernsteinFunction}(object)

\S4method{getLevyDensity}{InverseGaussianBernsteinFunction}(object)

\S4method{getStieltjesDensity}{InverseGaussianBernsteinFunction}(object)

\S4method{calcValue}{InverseGaussianBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{LinearBernsteinFunction}(.Object, scale)

\S4method{show}{LinearBernsteinFunction}(object)

\S4method{calcValue}{LinearBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{LinearBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...
)

\S4method{initialize}{ParetoBernsteinFunction}(.Object, alpha, x0)

\S4method{show}{ParetoBernsteinFunction}(object)

\S4method{getLevyDensity}{ParetoBernsteinFunction}(object)

\S4method{calcValue}{ParetoBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{PoissonBernsteinFunction}(.Object, eta)

\S4method{show}{PoissonBernsteinFunction}(object)

\S4method{getLevyDensity}{PoissonBernsteinFunction}(object)

\S4method{calcValue}{PoissonBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{initialize}{ScaledBernsteinFunction}(.Object, scale, original)

\S4method{show}{ScaledBernsteinFunction}(object)

\S4method{calcValue}{ScaledBernsteinFunction}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{ScaledBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...
)

\S4method{initialize}{SumOfBernsteinFunctions}(.Object, first, second)

\S4method{show}{SumOfBernsteinFunctions}(object)

\S4method{calcValue}{SumOfBernsteinFunctions}(object, x, cscale = 1, ...)

\S4method{calcIterativeDifference}{SumOfBernsteinFunctions}(
  object,
  x,
  difference_order = 0L,
  n = 1L,
  k = 0L,
  cscale = 1,
  ...
)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}

\item{d}{A positive integer, larger than two, for the \emph{dimension}.}

\item{cscale}{Positive number.}

\item{...}{pass-through parameter.}

\item{x}{A nonnegative numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{A nonnegative integer with the order of the
alternating iterated forward differences taken on the Bernstein function.}

\item{n, k}{Nonnegative numbers for the binomial factor.}

\item{method}{Method to calculate the result; use \code{method = "levy"} for
using the Lévy representation and \code{method = "stieltjes"} for using the
Stieltjes representation.}

\item{tolerance}{(Relative) tolerance, passed down to \code{\link[stats:integrate]{stats::integrate()}}.}

\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}

\item{original}{Derives from \linkS4class{BernsteinFunction}.}

\item{constant}{Nonnegative number.}

\item{coefficients}{Derives from
\linkS4class{ConvexCombinationOfBernsteinFunctions}.}

\item{points}{Derives from \linkS4class{ConvexCombinationOfBernsteinFunctions}.}

\item{lambda}{Positive number.}

\item{a}{Positive number.}

\item{eta}{Positive number.}

\item{scale}{Positive number.}

\item{x0}{Positive number.}

\item{first}{Derives from \linkS4class{BernsteinFunction}.}

\item{second}{Derives from \linkS4class{BernsteinFunction}.}
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
