% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{class}
\name{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction}
\title{Class for the \emph{Inverse Gaussian Bernstein function}}
\description{
For the inverse Gaussian Lévy subordinator with \eqn{\eta > 0},
the corresponding Bernstein function is the function
\deqn{
  \psi(x) = \sqrt{2x + \eta^2} - \eta, x>0.
}
}
\details{
For the inverse Gaussian Bernstein function, the higher-order alternating
iterated forward differences are not known in closed-form, but
we can use numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}})
to approximate it with the following representation:
\deqn{
 {(-1)}^{k-1} \Delta^{k} \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k \frac{1}{\sqrt{2\pi}
     u^{3/2}} e^{-\frac{1}{2}\eta^2 u} du, x>0, k>0.
}

This Bernstein function can be found on p. 309 in \insertCite{Mai2017a}{rmo}.
}
\section{Slots}{

\describe{
\item{\code{eta}}{The distribution parameter (drift of the
underlying Gaussian process)}
}}

\examples{
InverseGaussianBernsteinFunction()
InverseGaussianBernsteinFunction(eta=0.3)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}
}
