% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{InverseGaussianBernsteinFunction-constructor}
\alias{InverseGaussianBernsteinFunction-constructor}
\alias{initialize,InverseGaussianBernsteinFunction-method}
\alias{initialize,InverseGaussianBernsteinFunction,ANY-method}
\title{Constructor of \emph{inverse Gaussian Bernstein function}}
\usage{
\S4method{initialize}{InverseGaussianBernsteinFunction}(.Object, eta = 0.5)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{eta}{Non-negative number.}
}
\description{
Constructor of \emph{inverse Gaussian Bernstein function}
}
\examples{
InverseGaussianBernsteinFunction()
InverseGaussianBernsteinFunction(eta = 0.3)

}
\seealso{
\linkS4class{InverseGaussianBernsteinFunction}
}
