% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-LevyBernsteinFunction.R
\docType{class}
\name{LevyBernsteinFunction-class}
\alias{LevyBernsteinFunction-class}
\title{Virtual superclass for Bernstein functions with nonzero Lévy density}
\description{
A virtual superclass for all Bernstein functions which can represented
by a Lévy density (no drift or killing rate). That means that there exists
a Lévy measure \eqn{\nu} such that
\deqn{
  \psi(x) = \int_0^\infty (1 - e^{-ux}) \nu(du) , x > 0 .
}
}
\seealso{
\linkS4class{BernsteinFunction}, \code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Virtual Bernstein function classes}
