% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{LinearBernsteinFunction-constructor}
\alias{LinearBernsteinFunction-constructor}
\alias{initialize,LinearBernsteinFunction-method}
\alias{initialize,LinearBernsteinFunction,ANY-method}
\title{Constructor for \emph{linear Bernstein functions}}
\usage{
\S4method{initialize}{LinearBernsteinFunction}(.Object, scale = 1)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{scale}{Non-negative number.}
}
\description{
Constructor for \emph{linear Bernstein functions}
}
\examples{
LinearBernsteinFunction()
LinearBernsteinFunction(scale = 2)

}
\seealso{
\linkS4class{LinearBernsteinFunction}
}
