% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_rmo_longformat}
\alias{merge_rmo_longformat}
\title{Merges two dataframe in rmo long format into one}
\usage{
merge_rmo_longformat(x, y)
}
\arguments{
\item{x}{dataframe in rmo long format}

\item{y}{dataframe in rmo long format}
}
\value{
merged dataframe
}
\description{
Missing values in y will be replaced with the values from x.
}
