% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{ParetoBernsteinFunction-constructor}
\alias{ParetoBernsteinFunction-constructor}
\alias{initialize,ParetoBernsteinFunction-method}
\alias{initialize,ParetoBernsteinFunction,ANY-method}
\title{Constructor of the \emph{Pareto Bernstein function}}
\usage{
\S4method{initialize}{ParetoBernsteinFunction}(.Object, alpha = log2(2 - 0.5), x0 = 1e-04)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}

\item{x0}{Positive number.}
}
\description{
Constructor of the \emph{Pareto Bernstein function}
}
\examples{
ParetoBernsteinFunction()
ParetoBernsteinFunction(alpha = 0.2, x0 = 1)

}
\seealso{
\linkS4class{ParetoBernsteinFunction}
}
