% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{PoissonBernsteinFunction-constructor}
\alias{PoissonBernsteinFunction-constructor}
\alias{initialize,PoissonBernsteinFunction-method}
\alias{initialize,PoissonBernsteinFunction,ANY-method}
\title{Constructor of \emph{Poisson Bernstein function}}
\usage{
\S4method{initialize}{PoissonBernsteinFunction}(.Object, eta = 1, lambda = 1)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{eta}{Positive number.}

\item{lambda}{Positive number}
}
\description{
Constructor of \emph{Poisson Bernstein function}
}
\examples{
PoissonBernsteinFunction()
PoissonBernsteinFunction(lambda = 0.2, eta = 2)

}
\seealso{
\linkS4class{PoissonBernsteinFunction}
}
