% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rarmextmo_esm}
\alias{rarmextmo_esm}
\title{Sample from the armageddon ESM distribution}
\usage{
rarmextmo_esm(n, d, alpha, beta)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Dimension}

\item{alpha}{Shock intensity of individual shocks}

\item{beta}{Shock intensity of global shock}
}
\value{
\code{rarmextmo_esm} implements an optimized version of the
\emph{exogenous shock model} algorithm for the armageddon shock family and
returns an \eqn{n \times d}{n x d} array matrix with rows corresponding to
the independent samples of size \eqn{d}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate armageddon ESM distribution
with parameters \code{alpha} and \code{beta}.
}
\details{
\strong{Parameterisation}:
The armageddon ESM distribution is a special case of the exchangeable Marshall-Olkin
distribution where \eqn{\lambda_i = 0 \forall 1 < i < d}.
The individual shock rate is \eqn{\alpha = \lambda_1} and the global shock
rate is \eqn{\beta = \lambda_d}.
}
\examples{
rarmextmo_esm(10L, 2L, 0.5, 0.2)
rarmextmo_esm(10L, 2L, 0, 1)      ## comonotone
rarmextmo_esm(10L, 2L, 1, 0)      ## independence

}
\seealso{
\code{\link[=rmo]{rmo()}}

Other sampling-algorithms: 
\code{\link{rexmo}()},
\code{\link{rextmo_lfm}()},
\code{\link{rmo}()}
}
\concept{sampling-algorithms}
