% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-store.R
\name{rds_get_store_fn}
\alias{rds_get_store_fn}
\title{Generates the name of the rds file}
\usage{
rds_get_store_fn(year, zf, site, base_path)
}
\arguments{
\item{year}{year of data to save}

\item{zf}{zeitfenster of data to save}

\item{site}{of data to save}

\item{base_path}{relative to current directory or absolut path to rds store}
}
\value{
file name as string
}
\description{
The file name is based on the year, zeitfenster and site name. At the moment the design for the
store is to split the data in year, zeitfenster and site chunks and save every chunk as rds file.
}
\keyword{internal}
