% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-store.R
\name{rds_load_data}
\alias{rds_load_data}
\title{Load data from the rds store}
\usage{
rds_load_data(sites, zfs, years, parameters = NULL, base_path = "rds")
}
\arguments{
\item{sites}{character vector with the airmo_kurzname of all sites to load}

\item{zfs}{character vector with all zeitfenster to load}

\item{years}{integer vector with all years to load}

\item{parameters}{opt. character vector to only return the parameters in the vector.
Use NULL to get all parameters saved in the rds file. Default: NULL}

\item{base_path}{relative to current directory or absolut path to rds store}
}
\value{
dataframe in rmo long format
}
\description{
Loads the specified data from the rds store and returns it as dataframe in
rmo long format
}
\examples{
\dontrun{
load_data("Zch_Schimmelstrasse", "min30", 2017, c("PM10", "NOx", "NO2"), "tmp")
load_data(c("Zch_Schimmelstrasse", "Zch_Stampfenbachstrasse", "Zch_Heubeeribüel"),
          "d1", 2017, "O3_max_h1", "tmp")
}
}
