% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-store.R
\name{rds_merge_into_store}
\alias{rds_merge_into_store}
\title{Merges data into to the rds store}
\usage{
rds_merge_into_store(df, year, zf, site, base_path)
}
\arguments{
\item{df}{dataframe in rmo long format}

\item{year}{of data to save}

\item{zf}{zeitfenster of data to save}

\item{site}{of data to save}

\item{base_path}{relative to current directory or absolut path to rds store}
}
\description{
Merges data into to the rds store
}
\keyword{internal}
