% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds-store.R
\name{rds_read_store}
\alias{rds_read_store}
\title{Reads data from a rds file}
\usage{
rds_read_store(fn, parameters = NULL)
}
\arguments{
\item{fn}{file name of the rds file}

\item{parameters}{opt. character vector to only return the parameters in the vector.
Use NULL to get all parameters saved in the rds file. Default: NULL}
}
\value{
dataframe in rmo long format
}
\description{
Reads data from a rds file
}
\keyword{internal}
