% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rextmo_lfm}
\alias{rextmo_lfm}
\title{Sample with Lévy-frailty model with compound Poisson subordinator}
\usage{
rextmo_lfm(
  n,
  d,
  rate,
  rate_killing,
  rate_drift,
  rjump_name,
  rjump_arg_list = list()
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Dimension}

\item{rate}{Jump intensity of CPP subordinator}

\item{rate_killing}{Killing intensity of CPP subordinator}

\item{rate_drift}{Drift of CPP subordinator}

\item{rjump_name}{Name of jump sampling function for jumps of CPP
subordinator}

\item{rjump_arg_list}{A list with named arguments for jump sampling function
for jumps of CPP subordinator}
}
\value{
\code{rextmo_lfm} implements the Lévy-frailty model algorithm with a
compound Poisson subordinator and returns an \eqn{n \times d}{n x d} numeric
matrix with the rows corresponding to independent and identically
distributed samples of the corresponding \code{d}-variate extendible
Marshall-Olkin distribution.
}
\description{
Draws \code{n} independent samples from a \code{d}-variate extendible Marshall-Olkin
distribution corresponding to a LFM with a compound Poisson subordinator.
}
\details{
\strong{Model}:
The default times are defined as first exit times of a (killed) compound Poisson
subordinator \eqn{\Lambda} with \emph{killing intensity} \eqn{a}, \emph{drift} \eqn{b},
\emph{jump intensity} \eqn{\xi}, and \emph{jump distribution} \eqn{X}.

\strong{Model parameters}:
\itemize{
\item \code{rate} is the \emph{jump intensity} of the compound Poisson subordinator.
\item \code{rate_killing} is the \emph{killing intensity} of the compound Poisson
subordinator.
\item \code{rate_drift} is the \emph{drift} of the compound Poisson subordinator.
\item \code{rjump_name} and \code{rjump_arg_list} are the jump distribution name and
list with jump distribution parameters. Currently, \code{"rexp"}, \code{"rposval"},
and \code{"rpareto"} are possible.
}
}
\section{References}{
 For more information on this algorithm, see J.-F. Mai,
M. Scherer, "Simulating Copulas", World Scientific (2017), pp. 140 psqq.
}

\examples{
rextmo_lfm(10L, 2L, 0.5, 0.1, 0.2, "rposval", list("value"=1))
rextmo_lfm(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))
rextmo_lfm(10L, 2L, 0.5, 0, 0, "rpareto", list("alpha"=log2(2 - 0.5), x0 = 1e-4))

rextmo_lfm(10L, 2L, 0, 0, 1, "rposval", list("value"=1))  ## independence
rextmo_lfm(10L, 2L, 0, 1, 0, "rposval", list("value"=1))  ## comonotone

}
\seealso{
Other sampling-algorithms: 
\code{\link{rexmo_markovian}()},
\code{\link{rmo_esm_cuadras_auge}},
\code{\link{rmo_esm}()}
}
\concept{sampling-algorithms}
