% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo.R
\docType{package}
\name{rmo-package}
\alias{rmo-package}
\title{rmo: A package for the Marshall-Olkin distribution.}
\description{
The package contains fast implementations of sampling algorithms related to
the Marshall-Olkin distribution. It also contains an S4-class system for
Bernstein functions which can be used to create a large variety of
Marshall-Olkin parameterisations.
}
\section{Bernstein functions}{

A \emph{Bernstein function} is a non-negative function with non-negative
alternating iterative differences. These functions can be used to
generated parameterisations for the exchangeable Marshall-Olkin
distribution.
\itemize{
\item An object that derives from \linkS4class{BernsteinFunction} can be used to
generate the Marshall-Olkin intensities with \code{\link[=intensities]{intensities()}}.
It can be used to generate (scaled) exchangeable intensities with
\code{\link[=exIntensities]{exIntensities()}}.
\item Many families of Bernstein functions are available, e.g.
\linkS4class{ParetoBernsteinFunction}, \linkS4class{ExponentialBernsteinFunction}, and
\linkS4class{AlphaStableBernsteinFunction}.
\item Bernstein functions can be recombined by scaling, by summation or by
composition, which can be used to create new Bernstein functions with
\linkS4class{ScaledBernsteinFunction} and \linkS4class{SumOfBernsteinFunctions}.
}
}

\section{Sampling}{

Multiple sampling algorithms are provided. The semantic naming scheme \code{r*mo} is used, e.g.,
\code{\link[=rexmo]{rexmo()}} has \emph{(scaled) exchangeable shock-size arrival intensities} \code{ex_intensities} as input
and uses, as the default, the Markovian model for the death-counting process for sampling.
The following algorithms are implemented:
\itemize{
\item The \emph{exogenous shock model} and the \emph{Arnold model} for the general Marshall-Olkin distribution
are implemented in \code{\link[=rmo]{rmo()}}. The former should only be used in very small dimensions and the
latter can be used up until \eqn{d \leq 30}.
\item \emph{Markovian death-counting model} for exchangeable Marshall-Olkin distributions with \code{\link[=rexmo]{rexmo()}}.
\item \emph{Lévy-frailty model for compound Poisson subordinators} for the corresponding
subclass of extendible Marshall-Olkin distributions with \code{\link[=rextmo_lfm]{rextmo_lfm()}}.
\item \emph{Exogenous shock model} for the armageddon ESM subclass with \code{\link[=rarmextmo_esm]{rarmextmo_esm()}}.
}
}

