% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo_esm_cuadras_auge}
\alias{rmo_esm_cuadras_auge}
\alias{rcamo_esm}
\title{Sample from Cuadras-Auge distribution}
\usage{
rcamo_esm(n, d, alpha, beta)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Dimension}

\item{alpha}{Shock intensity of individual shocks}

\item{beta}{Shock intensity of global shock}
}
\value{
\code{rmo_esm_cuadras_auge} implements an optimized version of the
\emph{exogenous shock model} algorithm for the Cuadras-Augé family and
returns an \eqn{n \times d}{n x d} array matrix with rows corresponding to
the independent samples of size \eqn{d}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate Cuadras-Augé distribution
with parameters \code{alpha} and \code{beta}.
}
\details{
\strong{Parameterisation}:
The Cuadras-Augé distribution is a special case of the exchangeable Marshall-Olkin
distribution where \eqn{\lambda_i = 0 \forall 1 < i < d}.
The individual shock rate is \eqn{\alpha = \lambda_1} and the global shock
rate is \eqn{\beta = \lambda_d}.
}
\examples{
rcamo_esm(10L, 2L, 0.5, 0.2)
rcamo_esm(10L, 2L, 0, 1)      ## comonotone
rcamo_esm(10L, 2L, 1, 0)      ## independence

}
\seealso{
\code{\link[=rmo_esm]{rmo_esm()}}

Other sampling-algorithms: 
\code{\link{rexmo_markovian}()},
\code{\link{rextmo_lfm}()},
\code{\link{rmo_esm}()}
}
\concept{sampling-algorithms}
