% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{ScaledBernsteinFunction-constructor}
\alias{ScaledBernsteinFunction-constructor}
\alias{initialize,ScaledBernsteinFunction-method}
\alias{initialize,ScaledBernsteinFunction,ANY-method}
\title{Constructor for a \emph{scaled Bernstein function}}
\usage{
\S4method{initialize}{ScaledBernsteinFunction}(.Object, scale = 1, original = LinearBernsteinFunction())
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{scale}{Positive number.}

\item{original}{Derives from \linkS4class{BernsteinFunction}.}
}
\description{
Constructor for a \emph{scaled Bernstein function}
}
\examples{
ScaledBernsteinFunction()
original_bf <- AlphaStableBernsteinFunction(alpha = 0.5)
ScaledBernsteinFunction(scale = 2, original = original_bf)

}
\seealso{
\linkS4class{ScaledBernsteinFunction}
}
