% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-stieltjesDensity.R
\name{stieltjesDensity}
\alias{stieltjesDensity}
\title{Return the Stieltjes density of a Bernstein function}
\usage{
stieltjesDensity(object)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}
}
\description{
Returns the \emph{Stieltjes density} with \code{lower}, \code{upper}, and \code{type} attributes
if continuous and returns a \code{data.frame} with named columns \code{x} (atoms) and
\code{y} (weights) as well as a type attribute if discrete. The \code{type} attribute
is either \code{"continuous"} or \code{"discrete"}.
}
\seealso{
Other Bernstein function generics: 
\code{\link{exIntensities}()},
\code{\link{exQMatrix}()},
\code{\link{intensities}()},
\code{\link{levyDensity}()},
\code{\link{uexIntensities}()},
\code{\link{valueOf}()}
}
\concept{Bernstein function generics}
