% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allInitialize-S4.R
\docType{methods}
\name{SumOfBernsteinFunctions-constructor}
\alias{SumOfBernsteinFunctions-constructor}
\alias{initialize,SumOfBernsteinFunctions-method}
\alias{initialize,SumOfBernsteinFunctions,ANY-method}
\title{Constructor for a \emph{sum of Bernstein functions}}
\usage{
\S4method{initialize}{SumOfBernsteinFunctions}(
  .Object,
  first = ConstantBernsteinFunction(0.5),
  second = LinearBernsteinFunction(0.5)
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{first}{Derives from \linkS4class{BernsteinFunction}.}

\item{second}{Derives from \linkS4class{BernsteinFunction}.}
}
\description{
Constructor for a \emph{sum of Bernstein functions}
}
\examples{
SumOfBernsteinFunctions()
first_bf <- LinearBernsteinFunction(scale = 0.2)
second_bf <- AlphaStableBernsteinFunction(alpha = 0.5)
SumOfBernsteinFunctions(first = first_bf, second = second_bf)

}
\seealso{
\linkS4class{SumOfBernsteinFunctions}
}
