% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-uexIntensities.R
\name{uexIntensities}
\alias{uexIntensities}
\title{Calculate exchangeable shock-arrival intensities}
\usage{
uexIntensities(object, d, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{d}{a positive integer, larger than two, for the \emph{dimension}.}

\item{cscale}{a positive number for the \emph{composite scaling factor}.}

\item{...}{pass-through parameter}
}
\description{
Calculates (unscaled) \emph{exchangeable shock-arrival intensities}, see \code{\link[=rmo]{rmo()}}
and \code{\link[=rexmo]{rexmo()}}.
}
\seealso{
Other Bernstein function generics: 
\code{\link{exIntensities}()},
\code{\link{exQMatrix}()},
\code{\link{intensities}()},
\code{\link{levyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{valueOf}()}
}
\concept{Bernstein function generics}
