% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{methods}
\name{valueOf}
\alias{valueOf}
\alias{valueOf,LevyBernsteinFunction-method}
\alias{valueOf,LevyBernsteinFunction,ANY-method}
\alias{valueOf,CompleteBernsteinFunction-method}
\alias{valueOf,CompleteBernsteinFunction,ANY-method}
\alias{valueOf,LinearBernsteinFunction-method}
\alias{valueOf,LinearBernsteinFunction,ANY-method}
\alias{valueOf,ConstantBernsteinFunction-method}
\alias{valueOf,ConstantBernsteinFunction,ANY-method}
\alias{valueOf,ScaledBernsteinFunction-method}
\alias{valueOf,ScaledBernsteinFunction,ANY-method}
\alias{valueOf,SumOfBernsteinFunctions-method}
\alias{valueOf,SumOfBernsteinFunctions,ANY-method}
\alias{valueOf,PoissonBernsteinFunction-method}
\alias{valueOf,PoissonBernsteinFunction,ANY-method}
\alias{valueOf,AlphaStableBernsteinFunction-method}
\alias{valueOf,AlphaStableBernsteinFunction,ANY-method}
\alias{valueOf,InverseGaussianBernsteinFunction-method}
\alias{valueOf,InverseGaussianBernsteinFunction,ANY-method}
\alias{valueOf,ExponentialBernsteinFunction-method}
\alias{valueOf,ExponentialBernsteinFunction,ANY-method}
\alias{valueOf,GammaBernsteinFunction-method}
\alias{valueOf,GammaBernsteinFunction,ANY-method}
\alias{valueOf,ParetoBernsteinFunction-method}
\alias{valueOf,ParetoBernsteinFunction,ANY-method}
\title{Returns values for Bernstein functions}
\usage{
valueOf(object, x, difference_order, ...)

\S4method{valueOf}{LevyBernsteinFunction}(
  object,
  x,
  difference_order,
  ...,
  method = "levy",
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{CompleteBernsteinFunction}(
  object,
  x,
  difference_order,
  ...,
  method = c("stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{LinearBernsteinFunction}(object, x, difference_order = 0L, ...)

\S4method{valueOf}{ConstantBernsteinFunction}(object, x, difference_order = 0L, ...)

\S4method{valueOf}{ScaledBernsteinFunction}(object, x, difference_order = 0L, ...)

\S4method{valueOf}{SumOfBernsteinFunctions}(object, x, difference_order = 0L, ...)

\S4method{valueOf}{PoissonBernsteinFunction}(object, x, difference_order = 0L, ..., method = c("default", "levy"))

\S4method{valueOf}{AlphaStableBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "levy", "stieltjes"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{InverseGaussianBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{ExponentialBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{GammaBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)

\S4method{valueOf}{ParetoBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{object}{The Bernstein function object.}

\item{x}{The value at which the Bernstein function is to be evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{k} in
the representation).}

\item{...}{Further parameter (passed to \code{\link[stats:integrate]{stats::integrate()}})}

\item{method}{Method to calculate the result}

\item{tolerance}{(Relative) tolerance, passed down to \code{\link[stats:integrate]{stats::integrate()}}}
}
\description{
This method allows you to calculate the values for a Bernstein function and
its higher-order, alternating iterated forward differences, i.e.
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x), x > 0.
}
}
\seealso{
\linkS4class{BernsteinFunction}

\linkS4class{LevyBernsteinFunction}

\linkS4class{CompleteBernsteinFunction}

\linkS4class{LinearBernsteinFunction}

\linkS4class{ConstantBernsteinFunction}

\linkS4class{ScaledBernsteinFunction}

\linkS4class{SumOfBernsteinFunctions}

\linkS4class{PoissonBernsteinFunction}

\linkS4class{AlphaStableBernsteinFunction}

\linkS4class{InverseGaussianBernsteinFunction}

\linkS4class{ExponentialBernsteinFunction}

\linkS4class{GammaBernsteinFunction}

\linkS4class{ParetoBernsteinFunction}
}
