% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-valueOf.R
\name{valueOf}
\alias{valueOf}
\title{Evaluate Bernstein Functions}
\usage{
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{a nonnegativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{a nonnegative integer with the order of the
alternating iterated forward differences taken on the Bernstein function.}

\item{n, k}{nonnegative numbers for the binomial factor.}

\item{cscale}{a positive numeric scalar with the composite scaling factor.}

\item{...}{pass-through parameter.}
}
\description{
Calculate the values for a Bernstein function and its higher-order,
alternating iterated forward differences, possibly scaled by a binomial
coefficient, i.e.,
\deqn{
  {(-1)}^{j-1} \Delta^{j}{ \psi(c x) } ,
    \quad x > 0 .
}
The evaluation of Bernstein functions using this formula is usually not
numerically stable. Consequently, the various alternative approaches are used
dependent on the class of the Bernstein function.
}
\examples{
bf <- AlphaStableBernsteinFunction(alpha = 0.7)
valueOf(bf, 1:5)
}
\seealso{
Other Bernstein function generics: 
\code{\link{exIntensities}()},
\code{\link{exQMatrix}()},
\code{\link{intensities}()},
\code{\link{levyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{uexIntensities}()}
}
\concept{Bernstein function generics}
