% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-valueOf0.R
\name{valueOf0}
\alias{valueOf0}
\title{Evaluate Bernstein functions}
\usage{
valueOf0(object, x, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}

\item{x}{A nonnegative numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{cscale}{A positive number for the \emph{composite scaling factor}.}

\item{...}{Pass-through parameter.}
}
\description{
Calculate the values for a Bernstein function function.
}
\examples{
bf <- AlphaStableBernsteinFunction(alpha = 0.7)
valueOf0(bf, 1:5)
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{defaultMethod}()},
\code{\link{exQMatrix}()},
\code{\link{intensities}()},
\code{\link{levyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{valueOf}()}
}
\concept{Bernstein function generics}
