% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/levyDensity-S4.R, R/stieltjesDensity-S4.R, R/valueOf-S4.R
\docType{class}
\name{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction}
\alias{initialize,AlphaStableBernsteinFunction-method}
\alias{initialize,AlphaStableBernsteinFunction,ANY-method}
\alias{levyDensity,AlphaStableBernsteinFunction-method}
\alias{stieltjesDensity,AlphaStableBernsteinFunction-method}
\alias{valueOf,AlphaStableBernsteinFunction-method}
\alias{valueOf,AlphaStableBernsteinFunction,ANY-method}
\title{Class for the \emph{\eqn{\alpha}-stable Bernstein function}}
\usage{
\S4method{initialize}{AlphaStableBernsteinFunction}(.Object, alpha = log2(2 - 0.5))

\S4method{levyDensity}{AlphaStableBernsteinFunction}(object)

\S4method{stieltjesDensity}{AlphaStableBernsteinFunction}(object)

\S4method{valueOf}{AlphaStableBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "levy", "stieltjes"),
  tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{k} in
the representation).}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{method}{Method to calculate the result; use \code{method = "levy"} for
using the Lévy representation and \code{method = "stieltjes"} for using the
Stieltjes representation.}

\item{tolerance}{(Relative) tolerance, passed down to \code{\link[stats:integrate]{stats::integrate()}}}
}
\description{
For the \eqn{\alpha}-stable Lévy subordinator with \eqn{0 < \alpha < 1},
the corresponding Bernstein function is the power function with exponent
\eqn{\alpha}, i.e.
\deqn{
  \psi(x) = x^\alpha, \quad x>0.
}
}
\details{
For the \eqn{\alpha}-stable Bernstein function, the higher order alternating
iterated forward differences are known in closed form but cannot be evaluated
numerically without the danger of loss of significance. But we can use
numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k
     \alpha \frac{1}{\Gamma(1-\alpha) u^{1+\alpha}} du, x>0, k>0 .
}

This Bernstein function is no. 1 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize,AlphaStableBernsteinFunction-method}: Constructor

\item \code{levyDensity,AlphaStableBernsteinFunction-method}: see \linkS4class{LevyBernsteinFunction}

\item \code{stieltjesDensity,AlphaStableBernsteinFunction-method}: see \linkS4class{CompleteBernsteinFunction}.

\item \code{valueOf,AlphaStableBernsteinFunction-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}.
The default method for calculating the iterated differences uses the
Lévy representation.
}}

\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}.}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \frac{\alpha}{\Gamma(1-\alpha)} u^{-1 - \alpha} , \quad u > 0 .
}
}

\section{Stieltjes Density}{

\deqn{
  \sigma(du)
    = \frac{\sin(\alpha \pi)}{\pi} u^{\alpha - 1}, \quad u > 0 .
}
}

\examples{
AlphaStableBernsteinFunction()
AlphaStableBernsteinFunction(alpha = 0.5)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction},
'\link{AlphaStableBernsteinFunction-constructor}, \code{\link[=valueOf]{valueOf()}}
}
