% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-AlphaStableBernsteinFunction.R
\docType{class}
\name{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction}
\title{Class for the \eqn{\alpha}-stable Bernstein function}
\description{
For the \eqn{\alpha}-stable Lévy subordinator with \eqn{0 < \alpha < 1},
the corresponding Bernstein function is the power function with exponent
\eqn{\alpha}, i.e.
\deqn{
  \psi(x) = x^\alpha, \quad x>0.
}
}
\details{
For the \eqn{\alpha}-stable Bernstein function, the higher order alternating
iterated forward differences are known in closed form but cannot be evaluated
numerically without the danger of loss of significance. But we can use
numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k
     \alpha \frac{1}{\Gamma(1-\alpha) u^{1+\alpha}} du, x>0, k>0 .
}

This Bernstein function is no. 1 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
\subsection{Lévy density}{

\deqn{
  \nu(du)
    = \frac{\alpha}{\Gamma(1-\alpha)} u^{-1 - \alpha} , \quad u > 0 .
}
}

\subsection{Stieltjes Density}{

\deqn{
  \sigma(du)
    = \frac{\sin(\alpha \pi)}{\pi} u^{\alpha - 1}, \quad u > 0 .
}
}
}
\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}.}
}}

\examples{
AlphaStableBernsteinFunction()
AlphaStableBernsteinFunction(alpha = 0.5)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction},
\code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}

Other Complete Bernstein function classes: 
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}}

Other Algebraic Bernstein function classes: 
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}}
}
\concept{Algebraic Bernstein function classes}
\concept{Bernstein function classes}
\concept{Complete Bernstein function classes}
\concept{Levy Bernstein function classes}
