% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{class}
\name{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction}
\title{Class for the \emph{\eqn{\alpha}-stable Bernstein function}}
\description{
For the \eqn{\alpha}-stable Lévy subordinator with \eqn{0 < \alpha < 1},
the corresponding Bernstein function is the power function with exponent
\eqn{\alpha}, i.e.
\deqn{
  \psi(x) = x^\alpha, x>0.
}
}
\details{
For the \eqn{\alpha}-stable Bernstein function, the higher order alternating
iterated forward differences are known in closed form but cannot be evaluated
numerically without the danger of loss of significance. But we can use
numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k
     \alpha \frac{1}{\Gamma(1-\alpha) u^{1+\alpha}} du, x>0, k>0 .
}

This Bernstein function is no. 1 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}.}
}}

\examples{
bf <- AlphaStableBernsteinFunction(alpha=0.5)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}
}
