% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-AlphaStableBernsteinFunction.R
\docType{class}
\name{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction-class}
\alias{AlphaStableBernsteinFunction}
\alias{initialize,AlphaStableBernsteinFunction-method}
\alias{initialize,AlphaStableBernsteinFunction,ANY-method}
\alias{show,AlphaStableBernsteinFunction-method}
\alias{levyDensity,AlphaStableBernsteinFunction-method}
\alias{stieltjesDensity,AlphaStableBernsteinFunction-method}
\title{Class for the \eqn{\alpha}-stable Bernstein function}
\usage{
\S4method{initialize}{AlphaStableBernsteinFunction}(.Object, alpha)

\S4method{show}{AlphaStableBernsteinFunction}(object)

\S4method{levyDensity}{AlphaStableBernsteinFunction}(object)

\S4method{stieltjesDensity}{AlphaStableBernsteinFunction}(object)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}
}
\description{
For the \eqn{\alpha}-stable Lévy subordinator with \eqn{0 < \alpha < 1},
the corresponding Bernstein function is the power function with exponent
\eqn{\alpha}, i.e.
\deqn{
  \psi(x) = x^\alpha, \quad x>0.
}
}
\details{
For the \eqn{\alpha}-stable Bernstein function, the higher order alternating
iterated forward differences are known in closed form but cannot be evaluated
numerically without the danger of loss of significance. But we can use
numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k
     \alpha \frac{1}{\Gamma(1-\alpha) u^{1+\alpha}} du, x>0, k>0 .
}

This Bernstein function is no. 1 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize(AlphaStableBernsteinFunction)}: Constructor

\item \code{show(AlphaStableBernsteinFunction)}: Display the object.

\item \code{levyDensity(AlphaStableBernsteinFunction)}: see \linkS4class{LevyBernsteinFunction}

\item \code{stieltjesDensity(AlphaStableBernsteinFunction)}: see \linkS4class{CompleteBernsteinFunction}.

}}
\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}.}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \frac{\alpha}{\Gamma(1-\alpha)} u^{-1 - \alpha} , \quad u > 0 .
}
}

\section{Stieltjes Density}{

\deqn{
  \sigma(du)
    = \frac{\sin(\alpha \pi)}{\pi} u^{\alpha - 1}, \quad u > 0 .
}
}

\examples{
AlphaStableBernsteinFunction()
AlphaStableBernsteinFunction(alpha = 0.5)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
