% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BernsteinFunction.R
\docType{class}
\name{BernsteinFunction-class}
\alias{BernsteinFunction-class}
\title{Virtual superclass for Bernstein functions}
\description{
A virtual superclass for all implementations of the various classes of
\emph{Bernstein functions}.
}
\details{
A \emph{Bernstein function} is a nonnegative, nondecreasing, infinitely often
differentiable function with whose recursive finite forward differences have
alternating signs:
\deqn{
    {(-1)}^{i-1} \Delta^{i}{ \psi{(x)} }
        \geq 0 ,
            \quad \forall  i \in \mathbb{N}, x \geq 0 .
}
\subsection{Extendible Marshall--Olkin distributions}{

A Bernstein function an \emph{extendible Marshall–Olkin distribution} with
\emph{exchangeable shock-size arrival intensities}
\deqn{
  \eta_{i}
    = \binom{d}{i} {(-1)}^{i-1} \Delta^{i}{ \psi{(d-i)} } ,
         \quad 1 \leq i \leq d .
}

The (upper triagonal) infinitesimal Markov generator of the associated
death-counting process is calculated recursively:
\deqn{
  q_{0, i}^\ast
    = \eta_{i} ,
      \quad i \in {\{ 1 , \ldots , d \}} ,
}
and
\deqn{
  q_{i+1, j+1}^\ast
    = \frac{d-j}{d-i} q_{i,j}^\ast + \frac{j+1-i}{d-i} q_{i, j+1}^\ast ,
     \quad 0 \leq i < j \leq d .
}
}
}
\references{
\insertRef{Schilling2012a}{rmo}
}
\seealso{
\code{\link[=valueOf]{valueOf()}}, \code{\link[=intensities]{intensities()}}, \code{\link[=uexIntensities]{uexIntensities()}}, \code{\link[=exIntensities]{exIntensities()}},
\code{\link[=exQMatrix]{exQMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Virtual Bernstein function classes}
