% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-BernsteinFunction.R
\docType{class}
\name{BernsteinFunction-class}
\alias{BernsteinFunction-class}
\title{Virtual superclass for Bernstein function objects}
\description{
Bernstein functions parametrize extendible Marshall--Olkin distributions.
They are closed under addition, scalar multiplication, composite scalar
multiplication, and consequently also convex recombination:
\itemize{
\item Pass a Bernstein function object to \code{\link[=rextmo]{rextmo()}} to simulate from the
associated extendible MO distribution.
\item Use \linkS4class{SumOfBernsteinFunctions} for adding two Bernstein functions,
\item Use \linkS4class{ScaledBernsteinFunction} for scalar multiplication of a
Bernstein function,
\item Use \linkS4class{CompositeScaledBernsteinFunction} for composite scalar
multiplication of a Bernstein function, and
\item Use \linkS4class{ConvexCombinationOfBernsteinFunctions} for convex recombination
of Bernstein functions.
}
}
\details{
\subsection{Mathematical definition}{

A \emph{Bernstein function} is a nonnegative, nondecreasing, infinitely often
differentiable function whose recursive finite forward differences have
alternating signs:
\deqn{
    {(-1)}^{i-1} \Delta^{i}{ \psi{(x)} }
        \geq 0 ,
            \quad \forall  i \in \mathbb{N}, x \geq 0 .
}
}

\subsection{Lévy-Khintchine representation}{

Bernstein functions have the so-called Lévy-Khintchine representation:
\deqn{
   \psi{(x)}
       = a x + b
         + \int_{0}^{\infty}{
             {\left[ 1 - e^{-x y} \right]} {\nu{(dy)}}
           } ,
         \quad x \geq 0 ,
}
for some nonnegative constants \eqn{a} and \eqn{b} and a Lévy measure
\eqn{\nu}. A Lévy measure is a measure on the Borel sets of the nonnegative
real line that satisfies the following properties:
\deqn{
  \int_{0}^{\infty}{
    \min{\{ 1 , y \}} {\nu{(dy)}} < \infty .
  }
}
}

\subsection{Lévy-frailty model and ext. MO distributions}{

Bernstein functions are uniquely linked to extendible Marshall--Olkin
distributions via the Lévy-frailty model.

First, each Bernstein functions is uniquely linked to a Lévy subordinator via
the \emph{Lévy-Khintchine representation}:
\deqn{
  \Lambda{(t)}
    = \begin{cases}
      b t + \Lambda_{\nu}{(t)} , & \text{if } t < \epsilon , \\
      \infty , & \text{otherwise} ,
    \end{cases}
}
where \eqn{\Lambda_{\nu}{(t)}} is the Lévy subordinator associated with the
Lévy measure \eqn{\nu} and \eqn{\epsilon} is an independently exponentially
distributed random variable with rate \eqn{a}. If \eqn{\nu} is a finite
measure, the Lévy subordinator is a compound Poisson process with intensity
\eqn{\nu{(0, \infty)}} and jump sizes \eqn{\nu{(dy)} / \nu{((0, \infty))} }.

Second, the Lévy subordinator is unique linked to an
extendible Marshall--Olkin distribution via the stochastic representation:
\deqn{
  \tau_{i}
    = \inf{\left \{ t \geq 0 : \Lambda{(t)} > E_{i} \right \}} ,
      \quad 1 \leq i \leq d ,
}
for independently unit exponentially distributed random variables
\eqn{E_{i}}.
}

\subsection{Ext. MO parameter interpretation}{

The Lévy-frailty model motivates the following interpretation of the
parameters for ext. MO distributions:
\itemize{
\item \code{a} is the killing rate,
\item \code{b} is the drift,
\item \code{gamma} is a scaling factor for the total shock-arrival intensity,
\item \code{family} is the name of the pure-jump Lévy measure, and
\item \code{eta} are the pure-jump family parameters.
}

To understand the influence of these parameters on the extendible
Marshall–Olkin distribution's dependence properties, the following
considerations are helpful:
\itemize{
\item A pure-killing Bernstein function (i.e., \eqn{a > 0}, \eqn{b = 0}, and
\eqn{\nu \equiv 0}) corresponds to complete comonotonicity.
\item A pure-drift Bernstein function (i.e., \eqn{a = 0}, \eqn{b > 0}, and
\eqn{\nu \equiv 0}) corresponds to independence.
\item A pure-jump Bernstein function (i.e., \eqn{a = 0}, \eqn{b = 0}, and
\eqn{\nu \not\equiv 0}) can model various dependence structures. However,
larger jump intensities lead to weaker dependence and larger jump sizes
lead to stronger dependence.
}

Consequently, weighting these cases with the parameters \code{a}, \code{b}, and \code{gamma}
allows for a flexible modeling of the dependence structure.
}

\subsection{Ext.MO marginal and dependence properties}{

For a given Bernstein function, the marginal rate and lower-tail dependence
coefficient of the associated extendible Marshall–Olkin distribution can be
calculated using as follows:
\deqn{
  \text{Marginal rate}
    = \psi{(1)}
}
and
\deqn{
  \text{LTDC}
    = 2 - \psi{(2)} / \psi{(1)} .
}
}

\subsection{Evaluating Bernstein functions}{

In context of extendible Marshall--Olkin distributions, the following
expression is frequently evaluated:
\deqn{
  \binom{n}{k} {(-1)}^{j-1} \Delta^{j}{ \psi{( c x )} } ,
    \quad 0 \leq k \leq n , j \in \mathbb{N}, x \geq 0 .
}
The evaluation of Bernstein functions using this formula is usually not
numerically stable. Consequently, the various alternative approaches are used
dependent on the class of the Bernstein function. Use the method \code{\link[=valueOf]{valueOf()}}
to evaluate or approximate this expression for a given Bernstein function.
}

\subsection{Exchangeable Marshall–Olkin distributions}{

An alternative stochastic representation of an exchangeable Marshall–Olkin
distributions is given by the so-called \emph{Markovian death-counting model}. It
defines the components' death times as randomized order statistics simulated
via the Markovian death-counting processes with infinitesimal generator
matrix:
\deqn{
  q_{i, j}^\ast
    = \binom{d-i}{j-i} \begin{cases}
      -\psi{(d-i)} , & \text{if } i = j , \\
      {(-1)}^{j-i-1} \Delta^{j-i}{ \psi{(d-i)} } , & \text{if } i < j , \\
      0 , & \text{otherwise} .
   \end{cases}
}
The evaluation of the infinitesimal generator matrix using this formula is
usually not numerically stable. Consequently, the various alternative
approaches are used dependent on the class of the Bernstein function. Use the
method \code{\link[=exQMatrix]{exQMatrix()}} to evaluate or approximate this expression for a given
Bernstein function.

For the \emph{all-alive-state}, the generator's first row has the interpretation
of \emph{exchangeable shock-size-arrival intensities}:
\deqn{
  \eta_{i}
    = \binom{d}{i} {(-1)}^{i-1} \Delta^{i}{ \psi{(d-i)} } ,
         \quad 1 \leq i \leq d .
}
As noted above, their evaluation is usually not numerically stable, and
various alternative approaches are used dependent on the class of the
Bernstein function. Use the method \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}} to
evaluate or approximate them.
}

\subsection{The Exogenous shock model and the Arnold model}{

Another alternative stochastic representation of Marshall–Olkin distributions
is
}
}
\references{
\insertRef{Schilling2012a}{rmo}
\insertRef{Sloot2022a}{rmo}
}
\seealso{
\code{\link[=valueOf]{valueOf()}}, \code{\link[=intensities]{intensities()}}, \code{\link[=uexIntensities]{uexIntensities()}},
\code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}}, \code{\link[=exQMatrix]{exQMatrix()}}, \code{\link[=rextmo]{rextmo()}},
\code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Virtual Bernstein function classes}
