% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/valueOf-S4.R
\docType{class}
\name{BernsteinFunction-class}
\alias{BernsteinFunction-class}
\alias{valueOf}
\alias{uexIntensities}
\alias{intensities}
\alias{exIntensities}
\alias{exQMatrix}
\alias{uexIntensities,BernsteinFunction-method}
\alias{exIntensities,BernsteinFunction-method}
\alias{intensities,BernsteinFunction-method}
\alias{exQMatrix,BernsteinFunction-method}
\title{Virtual Class \code{BernsteinFunction} for Bernstein Functions}
\usage{
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

uexIntensities(object, d, ...)

intensities(object, d, ...)

exIntensities(object, d, ...)

exQMatrix(object, d, ...)

\S4method{uexIntensities}{BernsteinFunction}(object, d, ...)

\S4method{exIntensities}{BernsteinFunction}(object, d, ...)

\S4method{intensities}{BernsteinFunction}(object, d, ...)

\S4method{exQMatrix}{BernsteinFunction}(object, d, ...)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{j} in
the representation).}

\item{n, k}{Non-negative numbers for the binomial factors.}

\item{cscale}{Positive number, the composit scaling factor.}

\item{...}{Pass-through parameter}

\item{d}{Dimension}
}
\description{
A virtual superclass for all implementations of the various classes of
Bernstein functions which have a representaion
\deqn{
 \psi(x) = a + b x + \int_0^\infty (1 - e^{-ux}) \nu(du) , x > 0
}
for non-negative constants \eqn{a, b \geq 0}, called \emph{killing rate} and
\emph{drift}, and a \emph{Lévy measure} \eqn{\nu} on \eqn{(0, \infty)}.
}
\details{
Bernstein functions have a 1-1 relationship to Lévy subordinators s.t. for
every Bernstein function \eqn{\psi}, there exists a Lévy subordinator
\eqn{\Lambda} with
\deqn{
  E[e^{-x \Lambda_t}]
    = e^{-t \psi(x)} , t, x > 0 .
}
\itemize{
\item \eqn{a} Describes the (exponential) \emph{killing rate}, which
determines the rate of an exponentially distributed event that sends
the subordinator to infinity.
\item \eqn{b} Describes the linear \emph{drift} of the Lévy subordinator.
\item \eqn{\nu} determines the arrival-rate and shock-size distribution
for the jumps of the Lévy subordinator.
}

For a theoretic treatment of Bernstein function, we refer to
\insertCite{Schilling2012a}{rmo}.

The method \code{valueOf} is implemented different, depending on the specific
Bernstein function to get the best possible accuracy:
\itemize{
\item For \code{difference_order == 0L}, the values are calculated with closed
form formulas if possible.
\item For \emph{linear Bernstein functions} and \emph{constant Bernstein functions}
the iterated differences are calculated automatically.
\item For \emph{scaled Bernstein functions} and \emph{sums of Bernstein functions}
the result is calculted recursively.
\item For all other cases, the values are calculated using either the
Lévy-Khintchine or the Stieltjes representation. Which one is chosen,
depends on the case and can be manually overwritten with
\code{method = "levy"} or \code{method = "stieltjes"}.
}
}
\section{Functions}{
\itemize{
\item \code{valueOf}: Calculate the values for a compound scaled Bernstein function and its higher-order,
alternating iterated forward differences, possibly scaled by a binomial
\deqn{
    {(-1)}^{j-1} \Delta^j \psi(c x), x > 0.
  }

\item \code{uexIntensities}: Calculates unscaled \code{ex_intensities} parameter for \code{\link[=rexmo_mdcm]{rexmo_mdcm()}}.

\item \code{intensities}: Calculates \code{intensities} parameter for \code{\link[=rmo]{rmo()}}.

\item \code{exIntensities}: Calculates \code{ex_intensities} parameter for \code{\link[=rexmo_mdcm]{rexmo_mdcm()}}.

\item \code{exQMatrix}: Calculates the exchangeable Q-matrix implies by \code{\link[=rexmo_mdcm]{rexmo_mdcm()}}.
}}

\examples{
exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)

exIntensities(AlphaStableBernsteinFunction(4e-1), 3L)

intensities(AlphaStableBernsteinFunction(4e-1), 3L)

exQMatrix(AlphaStableBernsteinFunction(4e-1), 3L)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{LevyBernsteinFunction}, \linkS4class{CompleteBernsteinFunction},
\linkS4class{ConstantBernsteinFunction},
\linkS4class{ScaledBernsteinFunction},
\linkS4class{SumOfBernsteinFunctions},
\linkS4class{LinearBernsteinFunction},
\linkS4class{PoissonBernsteinFunction},
\linkS4class{AlphaStableBernsteinFunction},
\linkS4class{ExponentialBernsteinFunction},
\linkS4class{InverseGaussianBernsteinFunction},
\linkS4class{GammaBernsteinFunction},
\linkS4class{ParetoBernsteinFunction}
}
