% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{class}
\name{CompleteBernsteinFunction-class}
\alias{CompleteBernsteinFunction-class}
\title{Virtual Class \code{CompleteBernsteinFunction} for Complete Bernstein Functions}
\description{
A virtual superclass for all Bernstein functions which can representated
by a Stieltjes density (no drift or killing rate). That means that there exists
a Stieltjes measure \eqn{\sigma} such that
\deqn{
  \psi(x) = \int_0^\infty \frac{x}{x + u} \sigma(du) , x > 0 .
}
}
\seealso{
\linkS4class{LevyBernsteinFunction}
\linkS4class{BernsteinFunction}
}
