% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-CompleteBernsteinFunction.R
\docType{class}
\name{CompleteBernsteinFunction-class}
\alias{CompleteBernsteinFunction-class}
\title{Virtual superclass for complete Bernstein functions}
\description{
A virtual superclass for all Bernstein functions which can represented
by a Stieltjes density (no drift or killing rate). That means that there
exists a Stieltjes measure \eqn{\sigma} such that
\deqn{
  \psi(x) = \int_0^\infty \frac{x}{x + u} \sigma(du) , x > 0 .
}
}
\details{
\subsection{Evaluation of Complete Bernstein functions}{

For \emph{continuous Stieltjes densities}, the values of the Bernstein function
are calculated with \code{\link[stats:integrate]{stats::integrate()}} by using the representation
\deqn{
  \psi(x)
    = \int_{0}^{\infty} x \mathrm{Beta}(1, x + u) \sigma(du), \quad x > 0 ,
}
and the values of the iterated differences are calculated by using the
representation
\deqn{
  (-1)^{j-1} \Delta^{j} \psi(x)
    = \int_{0}^{\infty} u \mathrm{Beta}(j+1, x + u) \sigma(du) ,
    \quad x > 0 .
}

For \emph{discrete Lévy densities}
\eqn{\sigma(du) = \sum_{i} y_i \delta_{u_i}(du)},
the values of the Bernstein function are calculated by using the
representation
\deqn{
  \psi(x)
    = \sum_{i} x \mathrm{Beta}(1, x + u_i) y_i, \quad x > 0 ,
}
and the values of the iterated differences are calculated by using the
representation
\deqn{
  (-1)^{j-1} \Delta^{j} \psi(x)
    = \sum_{i} u_i \mathrm{Beta}(j+1, x + u_i) y_i ,
    \quad x > 0 .
}
}
}
\seealso{
\code{\link[=levyDensity]{levyDensity()}}, \code{\link[=stieltjesDensity]{stieltjesDensity()}}, \code{\link[=valueOf]{valueOf()}},
\code{\link[=intensities]{intensities()}}, \code{\link[=uexIntensities]{uexIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=exQMatrix]{exQMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{BernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Stieltjes Bernstein function classes}
\concept{Virtual Bernstein function classes}
