% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/allShow-S4.R, R/valueOf-S4.R
\docType{class}
\name{CompositeScaledBernsteinFunction-class}
\alias{CompositeScaledBernsteinFunction-class}
\alias{CompositeScaledBernsteinFunction}
\alias{initialize,CompositeScaledBernsteinFunction-method}
\alias{initialize,CompositeScaledBernsteinFunction,ANY-method}
\alias{show,CompositeScaledBernsteinFunction-method}
\alias{valueOf,CompositeScaledBernsteinFunction-method}
\alias{valueOf,CompositeScaledBernsteinFunction,ANY-method}
\title{Class for the \emph{composite scaled Bernstein function}}
\usage{
\S4method{initialize}{CompositeScaledBernsteinFunction}(.Object, cscale, original)

\S4method{show}{CompositeScaledBernsteinFunction}(object)

\S4method{valueOf}{CompositeScaledBernsteinFunction}(object, x, difference_order = 0L, n = 1, k = 0, cscale = 1, ...)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{cscale}{Positive number.}

\item{original}{Derives from \linkS4class{BernsteinFunction}.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{j} in
the representation).}

\item{n, k}{Non-negative numbers for the binomial factors.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
Bernstein functions are stable under composition, i.e. if \eqn{\psi} is
a Bernstein function and \code{c > 0}, then
\deqn{
  x \mapsto \psi(c x)
}
is also a Bernstein function.
}
\section{Functions}{
\itemize{
\item \code{initialize,CompositeScaledBernsteinFunction-method}: Constructor

\item \code{show,CompositeScaledBernsteinFunction-method}: Display the object.

\item \code{valueOf,CompositeScaledBernsteinFunction-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}
}}

\section{Slots}{

\describe{
\item{\code{cscale}}{The scale of the inner linear Bernstein function of the
composition.}

\item{\code{original}}{The original Bernstein function.}
}}

\examples{
CompositeScaledBernsteinFunction()
cscale <- 0.5
bf_original <- AlphaStableBernsteinFunction()
CompositeScaledBernsteinFunction(cscale = cscale, original = bf_original)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
