% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction}
\title{Class for the \emph{constant Bernstein function}}
\description{
A \emph{constant Bernstein function} is a Bernstein function with only a
constant part (for \eqn{x > 0}), i.e. \eqn{b = 0} and \eqn{\nu = 0}. In
particular,
\deqn{
  \psi(x) = a , x > 0
}
}
\section{Slots}{

\describe{
\item{\code{constant}}{The non-negative \emph{killing} parameter (i.e. \eqn{a}
in the representation)}
}}

\seealso{
\linkS4class{BernsteinFunction} \link{ConstantBernsteinFunction-constructor}
}
