% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ConstantBernsteinFunction.R
\docType{class}
\name{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction-class}
\alias{ConstantBernsteinFunction}
\title{Class for (almost) constant Bernstein functions}
\description{
A \emph{constant Bernstein function} is a Bernstein function with only a
constant part (for \eqn{x > 0}), i.e. \eqn{b = 0} and \eqn{\nu = 0}. In
particular,
\deqn{
  \psi(x) = a , x > 0
}
}
\section{Slots}{

\describe{
\item{\code{constant}}{The nonnegative \emph{killing} parameter (i.e. \eqn{a}
in the representation)}
}}

\examples{
ConstantBernsteinFunction()
ConstantBernsteinFunction(constant = 0.2)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Bernstein function boundary classes: 
\code{\link{LinearBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function boundary classes}
\concept{Bernstein function classes}
