% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ConvexCombinationOfBernsteinFunctions.R
\docType{class}
\name{ConvexCombinationOfBernsteinFunctions-class}
\alias{ConvexCombinationOfBernsteinFunctions-class}
\alias{ConvexCombinationOfBernsteinFunctions}
\title{Class for convex combinations of Bernstein functions}
\description{
Bernstein functions are stable under convex combinations, i.e. if
\eqn{\psi_1, \ldots, \psi_n} are Bernstein functions and
\eqn{c_1, \ldots, c_n > 0} are positive real values,
\deqn{
  x \mapsto c_1 \psi_1(x) + \cdots + c_n \psi_n(x) , x>0,
}
is also a Bernstein function.
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{Numeric vector of positive real values.}

\item{\code{points}}{List of Bernstein functions.}
}}

\examples{
# Create an object of class ConvexCombinationOfBernsteinFunctions
ConvexCombinationOfBernsteinFunctions()
ConvexCombinationOfBernsteinFunctions(
  coefficients = c(0.2, 0.5, 0.1),
  points = list(
    LinearBernsteinFunction(scale = 0.2),
    ConstantBernsteinFunction(constant = 0.5),
    AlphaStableBernsteinFunction(alpha = 0.5))
)
}
\seealso{
\code{\link[=valueOf]{valueOf()}}, \code{\link[=intensities]{intensities()}}, \code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}},
\code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}}, \code{\link[=exQMatrix]{exQMatrix()}}, \code{\link[=rextmo]{rextmo()}},
\code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Bernstein function transformer classes: 
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}
}
\concept{Bernstein function classes}
\concept{Bernstein function transformer classes}
