% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-exIntensities.R
\name{exIntensities}
\alias{exIntensities}
\title{Calculate exchangeable shock-size-arrival intensities}
\usage{
exIntensities(object, d, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{d}{a positive integer, larger than two, for the \emph{dimension}.}

\item{cscale}{a positive number for the \emph{composite scaling factor}.}

\item{...}{pass-through parameter}
}
\description{
Calculates \emph{exchangeable shock-size-arrival intensities}, the
\code{ex_intensities} parameter for \code{\link[=rexmo]{rexmo()}}.
}
\seealso{
Other Bernstein function generics: 
\code{\link{exQMatrix}()},
\code{\link{intensities}()},
\code{\link{levyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{uexIntensities}()},
\code{\link{valueOf}()}
}
\concept{Bernstein function generics}
