% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/levyDensity-S4.R, R/stieltjesDensity-S4.R, R/valueOf-S4.R
\docType{class}
\name{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction}
\alias{initialize,ExponentialBernsteinFunction-method}
\alias{initialize,ExponentialBernsteinFunction,ANY-method}
\alias{levyDensity,ExponentialBernsteinFunction-method}
\alias{stieltjesDensity,ExponentialBernsteinFunction-method}
\alias{valueOf,ExponentialBernsteinFunction-method}
\alias{valueOf,ExponentialBernsteinFunction,ANY-method}
\title{Class for the Exponential jump CPP Bernstein function}
\usage{
\S4method{initialize}{ExponentialBernsteinFunction}(.Object, lambda = 1)

\S4method{levyDensity}{ExponentialBernsteinFunction}(object)

\S4method{stieltjesDensity}{ExponentialBernsteinFunction}(object)

\S4method{valueOf}{ExponentialBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{lambda}{Positive number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{k} in
the representation).}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{method}{Method to calculate the result; use \code{method = "levy"} for
using the Lévy representation and \code{method = "stieltjes"} for using the
Stieltjes representation.}

\item{tolerance}{(Relative) tolerance, passed down to \code{\link[stats:integrate]{stats::integrate()}}}
}
\description{
For the Exponential jump CPP subordinator with \eqn{\lambda > 0},
the corresponding Bernstein function is
\deqn{
  \psi(x) = \frac{x}{x + \lambda}, x>0.
}
}
\details{
For the Exponential jump CPP Bernstein function, the higher order
alternating iterated forward differences are known in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \lambda \cdot B(k+1, x+\lambda), x>0, k>0 .
}

This Bernstein function is no. 4 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize,ExponentialBernsteinFunction-method}: Constructor

\item \code{levyDensity,ExponentialBernsteinFunction-method}: see \linkS4class{LevyBernsteinFunction}

\item \code{stieltjesDensity,ExponentialBernsteinFunction-method}: see \linkS4class{CompleteBernsteinFunction}

\item \code{valueOf,ExponentialBernsteinFunction-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}.
The default method for calculating the iterated differences uses the
Stieltjes representation.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The index \eqn{\lambda}.}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \lambda \operatorname{e}^{-\lambda u}, \quad u > 0 .
}
}

\section{Stieltjes density}{

\deqn{
  \sigma(du)
    = \delta_{\lambda}(du), \quad u > 0 .
}
}

\examples{
ExponentialBernsteinFunction()
ExponentialBernsteinFunction(lambda = 0.5)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction}, \link{ExponentialBernsteinFunction-constructor},
\code{\link[=valueOf]{valueOf()}}
}
