% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ExponentialBernsteinFunction.R
\docType{class}
\name{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction}
\alias{initialize,ExponentialBernsteinFunction-method}
\alias{initialize,ExponentialBernsteinFunction,ANY-method}
\alias{show,ExponentialBernsteinFunction-method}
\alias{levyDensity,ExponentialBernsteinFunction-method}
\alias{stieltjesDensity,ExponentialBernsteinFunction-method}
\title{Class for the Exponential jump CPP Bernstein function}
\usage{
\S4method{initialize}{ExponentialBernsteinFunction}(.Object, lambda)

\S4method{show}{ExponentialBernsteinFunction}(object)

\S4method{levyDensity}{ExponentialBernsteinFunction}(object)

\S4method{stieltjesDensity}{ExponentialBernsteinFunction}(object)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{lambda}{Positive number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}
}
\description{
For the Exponential jump CPP subordinator with \eqn{\lambda > 0},
the corresponding Bernstein function is
\deqn{
  \psi(x) = \frac{x}{x + \lambda}, x>0.
}
}
\details{
For the Exponential jump CPP Bernstein function, the higher order
alternating iterated forward differences are known in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \lambda \cdot B(k+1, x+\lambda), x>0, k>0 .
}

This Bernstein function is no. 4 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize,ExponentialBernsteinFunction-method}: Constructor

\item \code{show,ExponentialBernsteinFunction-method}: Display the object.

\item \code{levyDensity,ExponentialBernsteinFunction-method}: see \linkS4class{LevyBernsteinFunction}

\item \code{stieltjesDensity,ExponentialBernsteinFunction-method}: see \linkS4class{CompleteBernsteinFunction}
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The index \eqn{\lambda}.}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \lambda \operatorname{e}^{-\lambda u}, \quad u > 0 .
}
}

\section{Stieltjes density}{

\deqn{
  \sigma(du)
    = \delta_{\lambda}(du), \quad u > 0 .
}
}

\examples{
ExponentialBernsteinFunction()
ExponentialBernsteinFunction(lambda = 0.5)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\linkS4class{CompleteBernsteinFunction}, \code{\link[=valueOf]{valueOf()}}
}
