% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction-class}
\alias{ExponentialBernsteinFunction}
\title{Class for the Exponential jump CPP Bernstein function}
\description{
For the Exponential jump CPP subordinator with \eqn{\lambda > 0},
the corresponding Bernstein function is
\deqn{
  \psi(x) = \frac{x}{x + \lambda}, x>0.
}
}
\details{
For the Exponential jump CPP Bernstein function, the higher order
alternating iterated forward differences are known in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
   = \lambda \cdot B(k+1, x+\lambda), x>0, k>0 .
}

This Bernstein function is no. 4 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The index \eqn{\lambda}.}
}}

\examples{
ExponentialBernsteinFunction()
ExponentialBernsteinFunction(lambda=0.5)

}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}
}
