% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-GammaBernsteinFunction.R
\docType{class}
\name{GammaBernsteinFunction-class}
\alias{GammaBernsteinFunction-class}
\alias{GammaBernsteinFunction}
\title{Class for Gamma Bernstein functions}
\description{
The \emph{Gamma Bernstein function}, is the Bernstein function of a
subordinator with a (scaled) Gamma distribution. The representation is for
\eqn{a > 0}
\deqn{
  \psi(x) = \log(1 + \frac{x}{a}), x > 0.
}
}
\details{
For this Bernstein function, the higher-order alternating iterated forward
differences are known in closed form but cannot be evaluated numerically
without the danger of loss of significance. But we can use numerical
integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the
following representation:
\deqn{
  {(-1)}^{k-1} \Delta^{k} \psi(x)
    = \int_{0}^{\infty} e^{-ux} {(1 - e^{-u})}^{k}
      \frac{e^{-au}}{u} du, x>0, k>0.
}

This Bernstein function is no. 26 in the list of complete Bernstein functions
in Chp. 16 of \insertCite{Schilling2012a}{rmo}.

The Gamma Bernstein function has the Lévy density \eqn{\nu}:
\deqn{
  \nu(du)
    = \frac{\operatorname{e}^{-a u}}{u}, \quad u > 0 ,
}

and it has the Stieltjes density \eqn{\sigma}:
\deqn{
  \sigma(du)
    = u^{-1} du, u > a .
}
}
\section{Slots}{

\describe{
\item{\code{a}}{Scale parameter for the Lévy measure.}
}}

\examples{
# Create an object of class GammaBernsteinFunction
GammaBernsteinFunction()
GammaBernsteinFunction(a = 2)

# Create a Lévy density
bf <- GammaBernsteinFunction(a = 0.7)
levy_density <- levyDensity(bf)
integrate(
  function(x) pmin(1, x) * levy_density(x),
  lower = attr(levy_density, "lower"),
  upper = attr(levy_density, "upper")
)

# Create a Stieltjes density
bf <- GammaBernsteinFunction(a = 0.5)
stieltjes_density <- stieltjesDensity(bf)
integrate(
  function(x) 1/(1 + x) * stieltjes_density(x),
  lower = attr(stieltjes_density, "lower"),
  upper = attr(stieltjes_density, "upper")
)

# Evaluate the Bernstein function
bf <- GammaBernsteinFunction(a = 0.3)
valueOf(bf, 1:5)

# Calculate shock-arrival intensities
bf <- GammaBernsteinFunction(a = 0.8)
intensities(bf, 3)
intensities(bf, 3, method = "stieltjes")
intensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-arrival intensities
bf <- GammaBernsteinFunction(a = 0.4)
uexIntensities(bf, 3)
uexIntensities(bf, 3, method = "stieltjes")
uexIntensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-size arrival intensities
bf <- GammaBernsteinFunction(a = 0.2)
calcExShockSizeArrivalIntensities(bf, 3)
calcExShockSizeArrivalIntensities(bf, 3, method = "stieltjes")
calcExShockSizeArrivalIntensities(bf, 3, tolerance = 1e-4)

# Calculate the Markov generator
bf <- GammaBernsteinFunction(a = 0.6)
exQMatrix(bf, 3)
exQMatrix(bf, 3, method = "stieltjes")
exQMatrix(bf, 3, tolerance = 1e-4)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=levyDensity]{levyDensity()}}, \code{\link[=stieltjesDensity]{stieltjesDensity()}}, \code{\link[=valueOf]{valueOf()}},
\code{\link[=intensities]{intensities()}}, \code{\link[=uexIntensities]{uexIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=exQMatrix]{exQMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}

Other Complete Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Complete Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Logarithmic Bernstein function classes}
