% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-InverseGaussianBernsteinFunction.R
\docType{class}
\name{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction}
\alias{initialize,InverseGaussianBernsteinFunction-method}
\alias{initialize,InverseGaussianBernsteinFunction,ANY-method}
\alias{show,InverseGaussianBernsteinFunction-method}
\alias{levyDensity,InverseGaussianBernsteinFunction-method}
\alias{stieltjesDensity,InverseGaussianBernsteinFunction-method}
\title{Class for Inverse Gaussian Bernstein function}
\usage{
\S4method{initialize}{InverseGaussianBernsteinFunction}(.Object, eta)

\S4method{show}{InverseGaussianBernsteinFunction}(object)

\S4method{levyDensity}{InverseGaussianBernsteinFunction}(object)

\S4method{stieltjesDensity}{InverseGaussianBernsteinFunction}(object)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{eta}{Non-negative number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}
}
\description{
For the inverse Gaussian Lévy subordinator with \eqn{\eta > 0},
the corresponding Bernstein function is the function
\deqn{
  \psi(x) = \sqrt{2x + \eta^2} - \eta, x>0.
}
}
\details{
For the inverse Gaussian Bernstein function, the higher-order alternating
iterated forward differences are not known in closed-form, but
we can use numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}})
to approximate it with the following representation:
\deqn{
 {(-1)}^{k-1} \Delta^{k} \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k \frac{1}{\sqrt{2\pi}
     u^{3/2}} e^{-\frac{1}{2}\eta^2 u} du, x>0, k>0.
}

This Bernstein function can be found on p. 309 in \insertCite{Mai2017a}{rmo}.
Furthermore it is  a transformation of no. 2 in the list of complete Bernstein
functions in Chp. 16 of \insertCite{Schilling2012a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize(InverseGaussianBernsteinFunction)}: Constructor

\item \code{show(InverseGaussianBernsteinFunction)}: Display the object.

\item \code{levyDensity(InverseGaussianBernsteinFunction)}: see \linkS4class{LevyBernsteinFunction}

\item \code{stieltjesDensity(InverseGaussianBernsteinFunction)}: see \linkS4class{CompleteBernsteinFunction}

}}
\section{Slots}{

\describe{
\item{\code{eta}}{The distribution parameter (drift of the
underlying Gaussian process)}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \frac{1}{\sqrt{2 \pi u^3}} \operatorname{e}^{-\frac{1}{2} \eta^2 u} ,
    \quad u > 0 .
}
}

\section{Stieltjes density}{

\deqn{
  \sigma(du)
    = \frac{\sin(\pi / 2)}{\pi} \cdot \frac{\sqrt{2 x - \eta^2}}{x} , \quad u > \eta^2 / 2 .
}
}

\examples{
InverseGaussianBernsteinFunction()
InverseGaussianBernsteinFunction(eta = 0.3)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
