% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-InverseGaussianBernsteinFunction.R
\docType{class}
\name{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction-class}
\alias{InverseGaussianBernsteinFunction}
\title{Class for Inverse Gaussian Bernstein function}
\description{
For the inverse Gaussian Lévy subordinator with \eqn{\eta > 0},
the corresponding Bernstein function is the function
\deqn{
  \psi(x) = \sqrt{2x + \eta^2} - \eta, x>0.
}
}
\details{
For the inverse Gaussian Bernstein function, the higher-order alternating
iterated forward differences are not known in closed-form, but
we can use numerical integration (here: \code{\link[stats:integrate]{stats::integrate()}})
to approximate it with the following representation:
\deqn{
 {(-1)}^{k-1} \Delta^{k} \psi(x)
   = \int_0^\infty e^{-ux} (1-e^{-u})^k \frac{1}{\sqrt{2\pi}
     u^{3/2}} e^{-\frac{1}{2}\eta^2 u} du, x>0, k>0.
}

This Bernstein function can be found on p. 309 in \insertCite{Mai2017a}{rmo}.
Furthermore it is  a transformation of no. 2 in the list of complete
Bernstein functions in Chp. 16 of \insertCite{Schilling2012a}{rmo}.

The inverse Gaussian Bernstein function has the Lévy density \eqn{\nu}:
\deqn{
  \nu(du)
    = \frac{1}{\sqrt{2 \pi u^3}} \operatorname{e}^{-\frac{1}{2} \eta^2 u} ,
    \quad u > 0 ,
}

and it has the Stieltjes density \eqn{\sigma}:
\deqn{
  \sigma(du)
    = \frac{
        \sin(\pi / 2)
      }{
        \pi
      } \cdot \frac{
        \sqrt{2 x - \eta^2}
      }{
        x
      } ,
      \quad u > \eta^2 / 2 .
}
}
\section{Slots}{

\describe{
\item{\code{eta}}{The distribution parameter (drift of the
underlying Gaussian process).}
}}

\examples{
# Create an object of class InverseGaussianBernsteinFunction
InverseGaussianBernsteinFunction()
InverseGaussianBernsteinFunction(eta = 0.3)

# Create a Lévy density
bf <- InverseGaussianBernsteinFunction(eta = 0.7)
levy_density <- levyDensity(bf)
integrate(
  function(x) pmin(1, x) * levy_density(x),
  lower = attr(levy_density, "lower"),
  upper = attr(levy_density, "upper")
)

# Create a Stieltjes density
bf <- InverseGaussianBernsteinFunction(eta = 0.5)
stieltjes_density <- stieltjesDensity(bf)
integrate(
  function(x) 1/(1 + x) * stieltjes_density(x),
  lower = attr(stieltjes_density, "lower"),
  upper = attr(stieltjes_density, "upper")
)

# Evaluate the Bernstein function
bf <- InverseGaussianBernsteinFunction(eta = 0.3)
valueOf(bf, 1:5)

# Calculate shock-arrival intensities
bf <- InverseGaussianBernsteinFunction(eta = 0.8)
intensities(bf, 3)
intensities(bf, 3, method = "stieltjes")
intensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-arrival intensities
bf <- InverseGaussianBernsteinFunction(eta = 0.4)
calcExShockArrivalIntensities(bf, 3)
calcExShockArrivalIntensities(bf, 3, method = "stieltjes")
calcExShockArrivalIntensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-size arrival intensities
bf <- InverseGaussianBernsteinFunction(eta = 0.2)
calcExShockSizeArrivalIntensities(bf, 3)
calcExShockSizeArrivalIntensities(bf, 3, method = "stieltjes")
calcExShockSizeArrivalIntensities(bf, 3, tolerance = 1e-4)

# Calculate the Markov generator
bf <- InverseGaussianBernsteinFunction(eta = 0.6)
calcMDCMGeneratorMatrix(bf, 3)
calcMDCMGeneratorMatrix(bf, 3, method = "stieltjes")
calcMDCMGeneratorMatrix(bf, 3, tolerance = 1e-4)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=levyDensity]{levyDensity()}}, \code{\link[=stieltjesDensity]{stieltjesDensity()}}, \code{\link[=valueOf]{valueOf()}}, \code{\link[=intensities]{intensities()}},
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}

Other Complete Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}}

Other Algebraic Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}}
}
\concept{Algebraic Bernstein function classes}
\concept{Bernstein function classes}
\concept{Complete Bernstein function classes}
\concept{Levy Bernstein function classes}
