% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{LevyBernsteinFunction-class}
\alias{LevyBernsteinFunction-class}
\title{Virtual Class \code{LevyBernsteinFunction} for Levy Bernstein Functions}
\description{
A virtual superclass for all Bernstein functions which can representated
by a Lévy density (no drift or killing rate). That means that there exists
a Lévy measure \eqn{\nu} such that
\deqn{
  \psi(x) = \int_0^\infty (1 - e^{-ux}) \nu(du) , x > 0 .
}
}
\seealso{
\linkS4class{BernsteinFunction}
}
