% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-LevyBernsteinFunction.R
\docType{class}
\name{LevyBernsteinFunction-class}
\alias{LevyBernsteinFunction-class}
\title{Virtual superclass for Bernstein functions with nonzero Lévy density}
\description{
A virtual superclass for all Bernstein functions which can represented
by a Lévy density (no drift or killing rate). That means that there exists
a Lévy measure \eqn{\nu} such that
\deqn{
  \psi(x) = \int_0^\infty (1 - e^{-ux}) \nu(du) , x > 0 .
}
}
\details{
\subsection{Evaluation of Bernstein functions with Lévy densities}{

For \emph{continuous Lévy densities}, the values of the Bernstein function are
calculated with \code{\link[stats:integrate]{stats::integrate()}} by using the representation
\deqn{
  \psi(x)
    = \int_{0}^{\infty} (1 - \operatorname{e}^{-ux}) \nu(du), \quad x > 0 ,
}
and the values of the iterated differences are calculated by using the
representation
\deqn{
  (-1)^{j-1} \Delta^{j} \psi(x)
    = \int_{0}^{\infty}
      \operatorname{e}^{-ux} (1 - \operatorname{e}^{-u})^j \nu(du) ,
    \quad x > 0 .
}

For \emph{discrete Lévy densities} \eqn{\nu(du) = \sum_{i} y_i \delta_{u_i}(du)},
the values of the Bernstein function are calculated by using the
representation
\deqn{
  \psi(x)
    = \sum_{i} (1 - \operatorname{e}^{-u_i x}) y_i, \quad x > 0 ,
}
and the values of the iterated differences are calculated by using the
representation
\deqn{
  (-1)^{j-1} \Delta^{j} \psi(x)
    = \sum_{i} \operatorname{e}^{-u_i x} (1 - \operatorname{e}^{-u_i})^j y_i,
    \quad x > 0 .
}
}
}
\seealso{
\code{\link[=levyDensity]{levyDensity()}}, \code{\link[=valueOf]{valueOf()}}, \code{\link[=calcShockArrivalIntensities]{calcShockArrivalIntensities()}},
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Virtual Bernstein function classes: 
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
\concept{Virtual Bernstein function classes}
