% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{class}
\name{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction}
\title{Class for the \emph{linear Bernstein function}}
\description{
\emph{A linear Bernstein function} is a Bernstein function with only a drift,
i.e. \eqn{a = 0} and \eqn{\nu = 0}. In particular,
\deqn{
 \psi(x) = b x, x > 0.
}
}
\section{Slots}{

\describe{
\item{\code{scale}}{The non-negative \emph{drift} parameter
(i.e. \eqn{b} in the representation)}
}}

\examples{
LinearBernsteinFunction()
LinearBernsteinFunction(scale = 2)

}
\seealso{
\linkS4class{BernsteinFunction}
}
