% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-LinearBernsteinFunction.R
\docType{class}
\name{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction}
\alias{initialize,LinearBernsteinFunction-method}
\alias{initialize,LinearBernsteinFunction,ANY-method}
\alias{show,LinearBernsteinFunction-method}
\alias{valueOf,LinearBernsteinFunction-method}
\alias{valueOf,LinearBernsteinFunction,ANY-method}
\title{Class for the \emph{linear Bernstein function}}
\usage{
\S4method{initialize}{LinearBernsteinFunction}(.Object, scale)

\S4method{show}{LinearBernsteinFunction}(object)

\S4method{valueOf}{LinearBernsteinFunction}(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{scale}{Non-negative number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{a non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{a non-negative integer with the order of the alternating iterated
forward differences taken on the Bernstein function.}

\item{n, k}{non-negative numbers for the binomial factor.}

\item{cscale}{a positive numeric scalar with the composite scaling factor.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
\emph{A linear Bernstein function} is a Bernstein function with only a drift,
i.e. \eqn{a = 0} and \eqn{\nu = 0}. In particular,
\deqn{
 \psi(x) = b x, x > 0.
}
}
\section{Functions}{
\itemize{
\item \code{initialize,LinearBernsteinFunction-method}: Constructor

\item \code{show,LinearBernsteinFunction-method}: Display the object.

\item \code{valueOf,LinearBernsteinFunction-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}
}}

\section{Slots}{

\describe{
\item{\code{scale}}{The non-negative \emph{drift} parameter
(i.e. \eqn{b} in the representation)}
}}

\examples{
LinearBernsteinFunction()
LinearBernsteinFunction(scale = 2)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
