% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-LinearBernsteinFunction.R
\docType{class}
\name{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction-class}
\alias{LinearBernsteinFunction}
\title{Class for linear Bernstein functions}
\description{
\emph{A linear Bernstein function} is a Bernstein function with only a drift,
i.e., \eqn{a = 0} and \eqn{\nu = 0}. In particular,
\deqn{
 \psi(x) = b x, x > 0.
}
}
\section{Slots}{

\describe{
\item{\code{scale}}{The nonnegative \emph{drift} parameter (i.e. \eqn{b} in the
representation).}
}}

\examples{
# Create an object of class LinearBernsteinFunction
LinearBernsteinFunction()
LinearBernsteinFunction(scale = 0.2)

# Evaluate the Bernstein function
bf <- LinearBernsteinFunction(scale = 0.3)
calcIterativeDifference(bf, 1:5)

# Calculate shock-arrival intensities
bf <- LinearBernsteinFunction(scale = 0.8)
calcShockArrivalIntensities(bf, 3)

# Calculate exchangeable shock-arrival intensities
bf <- LinearBernsteinFunction(scale = 0.4)
calcExShockArrivalIntensities(bf, 3)

# Calculate exchangeable shock-size arrival intensities
bf <- LinearBernsteinFunction(scale = 0.2)
calcExShockSizeArrivalIntensities(bf, 3)

# Calculate the Markov generator
bf <- LinearBernsteinFunction(scale = 0.6)
calcMDCMGeneratorMatrix(bf, 3)
}
\seealso{
\code{\link[=calcIterativeDifference]{calcIterativeDifference()}}, \code{\link[=calcShockArrivalIntensities]{calcShockArrivalIntensities()}},
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Bernstein function boundary classes: 
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}
}
\concept{Bernstein function boundary classes}
\concept{Bernstein function classes}
