% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\alias{ex_intensities_constant}
\alias{intensities_constant}
\alias{ex_intensities_linear}
\alias{intensities_linear}
\alias{ex_intensities_armageddon}
\alias{intensities_armageddon}
\alias{ex_intensities_poisson}
\alias{intensities_poisson}
\alias{ex_intensities_alpha_stable}
\alias{intensities_alpha_stable}
\alias{ex_intensities_exponential}
\alias{intensities_exponential}
\alias{ex_intensities_gamma}
\alias{intensities_gamma}
\alias{ex_intensities_pareto}
\alias{intensities_pareto}
\alias{ex_intensities_inverse_gaussian}
\alias{intensities_inverse_gaussian}
\title{Convenience functions to MO parameters}
\usage{
ex_intensities_constant(d, constant)

intensities_constant(d, constant)

ex_intensities_linear(d, scale)

intensities_linear(d, scale)

ex_intensities_armageddon(d, alpha, beta)

intensities_armageddon(d, alpha, beta)

ex_intensities_poisson(d, eta)

intensities_poisson(d, eta)

ex_intensities_alpha_stable(d, alpha)

intensities_alpha_stable(d, alpha)

ex_intensities_exponential(d, lambda)

intensities_exponential(d, lambda)

ex_intensities_gamma(d, a)

intensities_gamma(d, a)

ex_intensities_pareto(d, alpha, x0)

intensities_pareto(d, alpha, x0)

ex_intensities_inverse_gaussian(d, eta)

intensities_inverse_gaussian(d, eta)
}
\arguments{
\item{d}{Dimension of the distribution}

\item{constant}{Rate for global shock, passed to \code{constant} parameter for
\linkS4class{ConstantBernsteinFunction}}

\item{scale}{Rates for individual shocks, passed to \code{scale} parameter for
\linkS4class{LinearBernsteinFunction}}

\item{alpha}{Alpha parameter for Pareto CPP, passed to \code{alpha}
parameter of \linkS4class{ParetoBernsteinFunction}}

\item{beta}{Rate for global shock, passed to \code{constant} parameter for
\linkS4class{ConstantBernsteinFunction}}

\item{eta}{\code{eta} parameter for the inverse Gaussian BF, passed to \code{eta}
parameter of \linkS4class{InverseGaussianBernsteinFunction}}

\item{lambda}{lambda parameter for Exponential-jump CPP subordinator, passed to \code{lambda}
parameter of \linkS4class{ExponentialBernsteinFunction}}

\item{a}{Shape parameter for Gamma subordinator, passed to \code{a} parameter of
\linkS4class{GammaBernsteinFunction}}

\item{x0}{Cutoff parameter for Pareto CPP, passed to \code{x0}
parameter of \linkS4class{ParetoBernsteinFunction}}
}
\description{
The following functions are convenience wrappers to create \code{intensities}
parameters for the MO distribution and \code{ex_intensities} parameters for the
exMO distribution.
}
\details{
For more information, please read the corresponding  documentation
for the respective Bernstein function.
}
\examples{
ex_intensities_constant(10L, constant = 1)

intensities_constant(10L, constant = 1)

ex_intensities_linear(10L, scale = 1)

intensities_linear(10L, scale = 1)

ex_intensities_armageddon(10L, alpha = 1, beta = 0.5)

intensities_armageddon(10L, alpha = 1, beta = 0.5)

ex_intensities_poisson(10L, eta = 0.2)

intensities_poisson(10L, eta = 0.2)

ex_intensities_alpha_stable(10L, alpha = 0.4)

intensities_alpha_stable(10L, alpha = 0.4)

ex_intensities_exponential(10L, lambda = 0.4)

intensities_exponential(10L, lambda = 0.4)

ex_intensities_gamma(10L, a = 0.3)

intensities_gamma(10L, a = 0.3)

ex_intensities_pareto(10L, alpha = 0.4, x0 = 5e-4)

intensities_pareto(10L, alpha = 0.4, x0 = 5e-4)

ex_intensities_inverse_gaussian(10L, eta = 0.4)

intensities_inverse_gaussian(10L, eta = 0.4)

}
\seealso{
\code{\link[=valueOf]{valueOf()}}

\linkS4class{ConstantBernsteinFunction}

\linkS4class{LinearBernsteinFunction}

\linkS4class{PoissonBernsteinFunction}

\linkS4class{AlphaStableBernsteinFunction}

\linkS4class{ExponentialBernsteinFunction}

\linkS4class{GammaBernsteinFunction}

\linkS4class{ParetoBernsteinFunction}

\linkS4class{InverseGaussianBernsteinFunction}
}
