% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/levyDensity-S4.R, R/valueOf-S4.R
\docType{class}
\name{ParetoBernsteinFunction-class}
\alias{ParetoBernsteinFunction-class}
\alias{ParetoBernsteinFunction}
\alias{initialize,ParetoBernsteinFunction-method}
\alias{initialize,ParetoBernsteinFunction,ANY-method}
\alias{levyDensity,ParetoBernsteinFunction-method}
\alias{valueOf,ParetoBernsteinFunction-method}
\alias{valueOf,ParetoBernsteinFunction,ANY-method}
\title{Class for the \emph{Pareto Bernstein function}}
\usage{
\S4method{initialize}{ParetoBernsteinFunction}(.Object, alpha = log2(2 - 0.5), x0 = 1e-04)

\S4method{levyDensity}{ParetoBernsteinFunction}(object)

\S4method{valueOf}{ParetoBernsteinFunction}(
  object,
  x,
  difference_order = 0L,
  ...,
  method = c("default", "stieltjes", "levy"),
  tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{alpha}{Positive number between zero and one (bounds excl.).}

\item{x0}{Positive number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{k} in
the representation).}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{method}{Method to calculate the result; use \code{method = "levy"} for
using the Lévy representation and \code{method = "stieltjes"} for using the
Stieltjes representation.}

\item{tolerance}{(Relative) tolerance, passed down to \code{\link[stats:integrate]{stats::integrate()}}}
}
\description{
For the Pareto-jump compound Poisson process with index \eqn{0 < \alpha < 1}
and cutoff point \eqn{x0}, the corresponding Bernstein function is
\deqn{
  \psi(x)
  = 1 - e^{-x x_0} + (x_0 x)^\alpha \Gamma(1-\alpha, x_0 x) ,
  x>0 .
}
}
\details{
For this Bernstein function, the higher-order alternating iterated forward
differences are known in closed form but cannot be evaluated numerically
without the danger of loss of significance. But we can use numerical
integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the following
representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
  = \int_{x_0}^\infty e^{-ux} (1-e^{-u})^k
  \alpha \frac{{x_0}^\alpha}{t^{1+\alpha}} du,
  x>0, k>0 .
}

The Pareto Bernstein function, in combination with a linear Bernstein
function can be used to approximate the Bernstein function of an
\eqn{\alpha}-stable Subordinator, see Sec. 5.3 of
\insertCite{Fernandez2015a}{rmo}.
}
\section{Functions}{
\itemize{
\item \code{initialize,ParetoBernsteinFunction-method}: Constructor

\item \code{levyDensity,ParetoBernsteinFunction-method}: see \linkS4class{LevyBernsteinFunction}

\item \code{valueOf,ParetoBernsteinFunction-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}.
The default method for calculating the iterated differences uses the
Lévy representation.
}}

\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}}

\item{\code{x0}}{The cutoff point \eqn{x_0}}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \alpha \frac{x_0^\alpha}{u^{\alpha + 1}}, \quad u > x_0 .
}
}

\examples{
ParetoBernsteinFunction()
ParetoBernsteinFunction(alpha = 0.2, x0 = 1)
}
\references{
\insertAllCited{}
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
