% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ParetoBernsteinFunction.R
\docType{class}
\name{ParetoBernsteinFunction-class}
\alias{ParetoBernsteinFunction-class}
\alias{ParetoBernsteinFunction}
\title{Class for Pareto Bernstein functions}
\description{
For the Pareto-jump compound Poisson process with index \eqn{0 < \alpha < 1}
and cutoff point \eqn{x0}, the corresponding Bernstein function is
\deqn{
  \psi(x)
  = 1 - e^{-x x_0} + (x_0 x)^\alpha \Gamma(1-\alpha, x_0 x) ,
  x>0 .
}
}
\details{
For this Bernstein function, the higher-order alternating iterated forward
differences are known in closed form but cannot be evaluated numerically
without the danger of loss of significance. But we can use numerical
integration (here: \code{\link[stats:integrate]{stats::integrate()}}) to approximate it with the following
representation:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x)
  = \int_{x_0}^\infty e^{-ux} (1-e^{-u})^k
  \alpha \frac{{x_0}^\alpha}{t^{1+\alpha}} du,
  x>0, k>0 .
}

The Pareto Bernstein function has the Lévy density \eqn{\nu}:
\deqn{
  \nu(du)
    = \alpha \frac{x_0^\alpha}{u^{\alpha + 1}}, \quad u > x_0 .
}

The Pareto Bernstein function, in combination with a linear Bernstein
function can be used to approximate the Bernstein function of an
\eqn{\alpha}-stable subordinator, see Sec. 5.3 of
\insertCite{Fernandez2015a}{rmo}.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{The index \eqn{\alpha}}

\item{\code{x0}}{The cutoff point \eqn{x_0}}
}}

\examples{
# Create an object of class ParetoBernsteinFunction
ParetoBernsteinFunction()
ParetoBernsteinFunction(alpha = 0.2, x0 = 1e-2)

# Create a Lévy density
bf <- ParetoBernsteinFunction(alpha = 0.7, x0 = 1e-2)
levy_density <- levyDensity(bf)
integrate(
  function(x) pmin(1, x) * levy_density(x),
  lower = attr(levy_density, "lower"),
  upper = attr(levy_density, "upper")
)

# Evaluate the Bernstein function
bf <- ParetoBernsteinFunction(alpha = 0.3, x0 = 1)
valueOf(bf, 1:5)

# Calculate shock-arrival intensities
bf <- ParetoBernsteinFunction(alpha = 0.8, x0 = 1e-2)
intensities(bf, 3)
intensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-arrival intensities
bf <- ParetoBernsteinFunction(alpha = 0.4, x0 = 1e-2)
calcExShockArrivalIntensities(bf, 3)
calcExShockArrivalIntensities(bf, 3, tolerance = 1e-4)

# Calculate exchangeable shock-size arrival intensities
bf <- ParetoBernsteinFunction(alpha = 0.2, x0 = 1e-2)
calcExShockSizeArrivalIntensities(bf, 3)
calcExShockSizeArrivalIntensities(bf, 3, tolerance = 1e-4)

# Calculate the Markov generator
bf <- ParetoBernsteinFunction(alpha = 0.6, x0 = 1e-2)
calcMDCMGeneratorMatrix(bf, 3)
calcMDCMGeneratorMatrix(bf, 3, tolerance = 1e-4)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=levyDensity]{levyDensity()}},  \code{\link[=valueOf]{valueOf()}}, \code{\link[=intensities]{intensities()}},
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}}

Other Algebraic Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}}
}
\concept{Algebraic Bernstein function classes}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
