% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction}
\title{Class for the \emph{Poisson Bernstein function}}
\description{
The Poisson process with arrival-rate \eqn{\lambda} and fixed jump size
\eqn{\eta} is a Lévy subordinator corresponding to the Bernstein function
\deqn{
  \psi(x) = \lambda (1 - e^{-x\eta}) , x>0.
}
}
\details{
For the Poisson Bernstein function, the higher-order alternatig iterated
foward differences can be calculated in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x) = e^{-u\eta} (1-e^{-\eta})^k, x>0, k>0.
}
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The (positive) arrival rate of the underlying Poisson process.}

\item{\code{eta}}{The fixed (positive) jump size of the Poisson process.}
}}

\seealso{
\linkS4class{BernsteinFunction} \linkS4class{LevyBernsteinFunction}
\link{PoissonBernsteinFunction-constructor}
}
