% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-PoissonBernsteinFunction.R
\docType{class}
\name{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction}
\title{Class for Poisson Bernstein functions}
\description{
The Poisson process with arrival-rate \eqn{\lambda} and fixed jump size
\eqn{\eta} is a Lévy subordinator corresponding to the Bernstein function
\deqn{
  \psi(x) = 1 - e^{-x\eta}, x>0.
}
}
\details{
For the Poisson Bernstein function, the higher-order alternating iterated
forward differences can be calculated in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x) = e^{-u\eta} (1-e^{-\eta})^k, x>0, k>0.
}
\subsection{Lévy density}{

\deqn{
  \nu(du)
    = \lambda \delta_{\eta}(du), \quad u > 0 .
}
}
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The (positive) arrival rate of the underlying Poisson process.}

\item{\code{eta}}{The fixed (positive) jump size of the Poisson process.}
}}

\examples{
PoissonBernsteinFunction()
PoissonBernsteinFunction(eta = 2)
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction}
\code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Levy Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}}

Other Bernstein function boundary classes: 
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}}
}
\concept{Bernstein function boundary classes}
\concept{Bernstein function classes}
\concept{Levy Bernstein function classes}
