% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-PoissonBernsteinFunction.R
\docType{class}
\name{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction-class}
\alias{PoissonBernsteinFunction}
\alias{initialize,PoissonBernsteinFunction-method}
\alias{initialize,PoissonBernsteinFunction,ANY-method}
\alias{show,PoissonBernsteinFunction-method}
\alias{levyDensity,PoissonBernsteinFunction-method}
\title{Class for Poisson Bernstein functions}
\usage{
\S4method{initialize}{PoissonBernsteinFunction}(.Object, eta)

\S4method{show}{PoissonBernsteinFunction}(object)

\S4method{levyDensity}{PoissonBernsteinFunction}(object)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{eta}{Positive number.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}
}
\description{
The Poisson process with arrival-rate \eqn{\lambda} and fixed jump size
\eqn{\eta} is a Lévy subordinator corresponding to the Bernstein function
\deqn{
  \psi(x) = 1 - e^{-x\eta}, x>0.
}
}
\details{
For the Poisson Bernstein function, the higher-order alternating iterated
forward differences can be calculated in closed form:
\deqn{
  {(-1)}^{k-1} \Delta^k \psi(x) = e^{-u\eta} (1-e^{-\eta})^k, x>0, k>0.
}
}
\section{Functions}{
\itemize{
\item \code{initialize(PoissonBernsteinFunction)}: Constructor

\item \code{show(PoissonBernsteinFunction)}: Display the object.

\item \code{levyDensity(PoissonBernsteinFunction)}: see \linkS4class{LevyBernsteinFunction}

}}
\section{Slots}{

\describe{
\item{\code{lambda}}{The (positive) arrival rate of the underlying Poisson process.}

\item{\code{eta}}{The fixed (positive) jump size of the Poisson process.}
}}

\section{Lévy density}{

\deqn{
  \nu(du)
    = \lambda \delta_{\eta}(du), \quad u > 0 .
}
}

\examples{
PoissonBernsteinFunction()
PoissonBernsteinFunction(eta = 2)
}
\seealso{
\linkS4class{BernsteinFunction}, \linkS4class{LevyBernsteinFunction}
\code{\link[=valueOf]{valueOf()}}
}
