% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-airmo.R
\name{read_airmo_csv}
\alias{read_airmo_csv}
\title{Reads an AIRMO Export in CSV Format}
\usage{
read_airmo_csv(fn)
}
\arguments{
\item{fn}{CSV file name, can be relative to current directory or absolut}
}
\value{
dataframe in rmo long format
}
\description{
Reads the data from a AIRMO CSV file and returns it as a dataframe in rmo long format
}
