% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-airmo.R
\name{read_airmo_dat}
\alias{read_airmo_dat}
\title{Reads an AIRMO Export in DAT Format}
\usage{
read_airmo_dat(fn)
}
\arguments{
\item{fn}{DAT file name, can be relative to current directory or absolut}
}
\value{
dataframe in rmo long format
}
\description{
Reads the data from a AIRMO DAT file and returns it as a dataframe in rmo long format
}
