% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rexmo}
\alias{rexmo}
\title{Sample from an exchangeable Marshall–Olkin distribution}
\usage{
rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))
}
\arguments{
\item{n}{an integer for the \emph{number of samples}.}

\item{d}{an integer for the \emph{dimension}.}

\item{ex_intensities}{a numeric vector with the exchangeable \emph{shock-size arrival intensities}.}

\item{method}{a string indicating which sampling algorithm should be used.
Use "MDCM" for the \emph{Markovian death-counting model}, "AM" for the \emph{Arnold model},
and "ESM" for the \emph{exogenous shock model}.}
}
\value{
\code{rexmo} returns a numeric matrix with \eqn{n} rows and \eqn{d} columns with the rows
corresponding to iid distributed samples of a \eqn{d}-variate \emph{exchangeable Marshall–Olkin
distribution} with exchangeable \emph{shock-size arrival intensities} \code{ex_intensities}.
}
\description{
Draws \eqn{n} iid samples from a \eqn{d}-variate \emph{exchangeable Marshall–Olkin
distribution} parametrized by exchangeable \emph{shock-size arrival rates}.
}
\section{Parametrization}{
The exchangeable Marshall–Olkin distribution has the following property:
a \emph{shock arrival intensity} \eqn{\lambda_I} for a set \eqn{I} only depends on the cardinality of
\eqn{I}, i.e., the \emph{exchangeable shock arrival intensities} are defined for \eqn{I} with \eqn{i =
\lvert I\rvert} by \eqn{\lambda_i = \lambda_I}, see \insertCite{@see Sec. 3.2 @Mai2017a}{rmo}.
The \emph{exchangeable Marshall–Olkin distribution} is parametrized by the exchangeable \emph{shock-size
arrival intensities} defined by
\deqn{
 \eta_i
   = \binom{d}{i} \lambda_i ,
   \quad i \in {\{ 1, \ldots, d \}} .
}
}

\section{Simulation algorithms}{
\subsection{General Marshall–Olkin sampling algorithms}{

The \emph{ESM} and \emph{AM} simulation algorithms for the general \emph{Marshall–Olkin distribution} can be
used; for this, the exchangeable \emph{shock-size arrival intensities} are converted to the
corresponding \emph{shock arrival intensities} and passed to \code{\link[=rmo]{rmo()}}.
}

\subsection{Markovian death-set model}{

TBD; see \insertCite{@see pp. 122 psqq. @Mai2017a}{rmo} for a similar algorithm.
}
}

\examples{
rexmo(10, 3, c(1.2, 0.3, 0.4))
rexmo(10, 3, c(3, 0, 0))          ## independence
rexmo(10, 3, c(0, 0, 1))          ## comonotone

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "AM")
rexmo(10, 3, c(3, 0, 0), method = "AM")          ## independence
rexmo(10, 3, c(0, 0, 1), method = "AM")          ## comonotone

rexmo(10, 3, c(1.2, 0.3, 0.4), method = "ESM")
rexmo(10, 3, c(3, 0, 0), method = "ESM")          ## independence
rexmo(10, 3, c(0, 0, 1), method = "ESM")          ## comonotone

}
\references{
\insertAllCited{}
}
\seealso{
Other sampling-algorithms: 
\code{\link{rextmo}()},
\code{\link{rmo}()},
\code{\link{rpextmo}()}
}
\concept{sampling-algorithms}
