% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rexmo}
\alias{rexmo}
\title{Sample from an exchangeable Marshall-Olkin distribution}
\usage{
rexmo(n, d, ex_intensities, method = c("MDCM", "AM", "ESM"))
}
\arguments{
\item{n}{an integer for the number of samples.}

\item{d}{an integer for the dimension of the sample.}

\item{ex_intensities}{a numeric vector with the exchangeable Marshall-Olkin exchangeable
shock-size arrival intensities.}

\item{method}{a character vector indicating which sampling algorithm should be used.
Use "MDCM" for the \emph{Markovian death-counting model}, "AM" for the \emph{Arnold model},
and "ESM" for the \emph{exogenous shock model}.}
}
\value{
\code{rexmo} implements the Markovian model for the death-counting process of the exchangeable
subclass and returns an  \eqn{n \times d}{n x d} numeric matrix with the rows corresponding to
independent and identically distributed samples of a \eqn{d} variate exchangeable Marshall-Olkin
distribution with exchangeable shock-size arrival intensities \code{ex_intensities}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate exchangeable Marshall-Olkin
distribution with shock-size arrival rates \code{ex_intensities}.
}
\details{
\strong{Parameterisation}:
The exchangeable Marshall--Olkin distribution has the property that \eqn{\lambda_I} only depends
on the cardinality of \eqn{I}, i.e., the \emph{(unscaled) exchangeable shock arrival  intensities} are
defined for \eqn{I} with \eqn{i = \lvert I\rvert} by \eqn{\lambda_i = \lambda_I}, see \code{\link[=rmo]{rmo()}}.
The \emph{(scaled) exchangeable shock-size arrival intensities} are defined by
\eqn{\eta_i = \binom{d}{i} \lambda_i}.
}
\section{References}{

For more information on a similar algorithm, see J.-F. Mai, M. Scherer,
"Simulating Copulas", World Scientific (2017), pp. 122 psqq.
}

\examples{
rexmo(10, 2, c(2 * 0.4, 0.2))
rexmo(10, 2, c(2, 0))          ## independence
rexmo(10, 2, c(0, 1))          ## comonotone

rexmo(10, 2, c(2 * 0.4, 0.2), method = "AM")
rexmo(10, 2, c(2, 0), method = "AM")          ## independence
rexmo(10, 2, c(0, 1), method = "AM")          ## comonotone

rexmo(10, 2, c(2 * 0.4, 0.2), method = "ESM")
rexmo(10, 2, c(2, 0), method = "ESM")          ## independence
rexmo(10, 2, c(0, 1), method = "ESM")          ## comonotone

}
\seealso{
Other sampling-algorithms: 
\code{\link{rextmo}()},
\code{\link{rmo}()},
\code{\link{rpextmo}()}
}
\concept{sampling-algorithms}
