% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-rextmo.R
\name{rextmo}
\alias{rextmo}
\title{Simulate from extendible Marshall–Olkin distributions}
\usage{
rextmo(n, d, bf, method = c("MDCM", "AM", "ESM"))
}
\arguments{
\item{n}{An integer for the \emph{number of samples}.}

\item{d}{An integer for the \emph{dimension}.}

\item{bf}{A \linkS4class{BernsteinFunction} representing the \emph{Bernstein function}
of a \emph{extendible Marshall–Olkin distribution}.}

\item{method}{A string indicating which sampling algorithm should be used.
Use \code{"MDCM"} for the \emph{Markovian death-counting model}, \code{"AM"} for the
\emph{Arnold model}, and \code{"ESM"} for the \emph{exogenous shock model}. We recommend
using the \emph{ESM} only for small dimensions; the \emph{AM} can be used up until
dimension \eqn{30}.}
}
\value{
\code{rextmo} returns a numeric matrix of size \code{n} x \code{d} rows and \code{d}. Each row
corresponds to an independently and identically (iid) distributed sample from
a \code{d}-variate \emph{extendible Marshall–Olkin distribution} with specified
parameters.
}
\description{
Draws \code{n} iid samples from a \code{d}-variate \emph{extendible Marshall–Olkin
distribution} parametrized by a Bernstein function \code{bf}.
}
\details{
The \emph{extendible Marshall–Olkin distribution} has the survival function
\deqn{
    \bar{F}{(t)}
        = \exp{\left\{
            - \sum_{i=1}^{d}{ {[ \psi{(i)} - \psi{(i-1)} ]} t_{[i]} }
          \right\}} ,
            \quad t = {(t_{1}, \ldots, t_{d})} > 0 ,
}
for \emph{Bernstein functions} \eqn{\psi}, see \linkS4class{BernsteinFunction}, and
\eqn{t_{[1]} \geq \cdots \geq t_{[d]}}, see \insertCite{Mai2017a}{rmo}.

The relationship between \emph{Bernstein functions} and
\emph{exchangeable shock-size arrival intensities} of the
\emph{exchangeable Marshall–Olkin distribution}, see \code{\link[=rexmo]{rexmo()}}, is as follows:
\deqn{
    \eta_{i}
        = \binom{d}{i} {(-1)}^{i-1} \Delta{ \psi{(d-i)} } ,
            \quad i \in {\{ 1 , \ldots , d \}} .
}

This formula for the \emph{exchangeable shock-size arrival intensities} is not
numerically stable in higher dimensions, and \code{\link[=rextmo]{rextmo()}} uses
approximation techniques from \insertCite{Sloot2022a}{rmo} to calculate them.
}
\examples{
rextmo(
  10, 3,
  AlphaStableBernsteinFunction(alpha = log2(2 - 0.5))
)
# independence
rextmo(
  10, 3,
  LinearBernsteinFunction(scale = 1)
)
# comonotone
rextmo(
  10, 3,
  ConstantBernsteinFunction(constant = 1)
)

rextmo(
  10, 3,
  AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
  method = "AM"
)
# independence
rextmo(
  10, 3,
  LinearBernsteinFunction(scale = 1),
  method = "AM"
)
# comonotone
rextmo(
  10, 3,
  ConstantBernsteinFunction(constant = 1),
  method = "AM"
)

rextmo(
  10, 3,
  AlphaStableBernsteinFunction(alpha = log2(2 - 0.5)),
  method = "ESM"
)
# independence
rextmo(
  10, 3,
  LinearBernsteinFunction(scale = 1),
  method = "ESM"
)
# comonotone
rextmo(
  10, 3,
  ConstantBernsteinFunction(constant = 1),
  method = "ESM"
)
}
\references{
\insertAllCited{}
}
\seealso{
Other sampling-algorithms: 
\code{\link{rexmo}()},
\code{\link{rmo}()},
\code{\link{rpextmo}()}
}
\concept{sampling-algorithms}
