% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo-deprecated.R
\name{rmo-deprecated}
\alias{rmo-deprecated}
\alias{exIntensities}
\alias{exIntensities-deprecated}
\alias{exIntensities,BernsteinFunction-method}
\alias{uexIntensities}
\alias{uexIntensities-deprecated}
\alias{uexIntensities,BernsteinFunction-method}
\alias{exQMatrix}
\alias{exQMatrix-deprecated}
\alias{exQMatrix,BernsteinFunction-method}
\alias{intensities}
\alias{intensities-deprecated}
\alias{intensities,BernsteinFunction-method}
\alias{defaultMethod}
\alias{defaultMethod-deprecated}
\alias{defaultMethod,BernsteinFunction-method}
\alias{levyDensity}
\alias{levyDensity-deprecated}
\alias{levyDensity,BernsteinFunction-method}
\alias{stieltjesDensity}
\alias{stieltjesDensity-deprecated}
\alias{stieltjesDensity,BernsteinFunction-method}
\alias{valueOf}
\alias{valueOf-deprecated}
\alias{valueOf,BernsteinFunction-method}
\alias{valueOf0}
\alias{valueOf0-deprecated}
\alias{valueOf0,BernsteinFunction-method}
\title{Deprecated functions in rmo}
\usage{
exIntensities(object, d, cscale = 1, ...)

uexIntensities(object, d, cscale = 1, ...)

exQMatrix(object, d, cscale = 1, ...)

intensities(object, d, cscale = 1, ...)

defaultMethod(object)

levyDensity(object)

stieltjesDensity(object)

valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)

valueOf0(object, x, cscale = 1, ...)
}
\description{
These functions are deprecated and will be removed in future versions of rmo.
}
\details{
\itemize{
\item \code{\link[=exIntensities]{exIntensities()}} is deprecated in favor of
\code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}}.
\item \code{\link[=uexIntensities]{uexIntensities()}} is deprecated in favor of
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}.
\item \code{\link[=exQMatrix]{exQMatrix()}} is deprecated in favor of \code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}.
\item \code{\link[=intensities]{intensities()}} is deprecated in favor of \code{\link[=calcShockArrivalIntensities]{calcShockArrivalIntensities()}}.
\item \code{\link[=defaultMethod]{defaultMethod()}} is deprecated in favor of \code{\link[=getDefaultMethodString]{getDefaultMethodString()}}.
\item \code{\link[=levyDensity]{levyDensity()}} is deprecated in favor of \code{\link[=getLevyDensity]{getLevyDensity()}}.
\item \code{\link[=stieltjesDensity]{stieltjesDensity()}} is deprecated in favor of \code{\link[=getStieltjesDensity]{getStieltjesDensity()}}.
\item \code{\link[=valueOf]{valueOf()}} is deprecated in favor of \code{\link[=calcIterativeDifference]{calcIterativeDifference()}}.
\item \code{\link[=valueOf0]{valueOf0()}} is deprecated in favor of \code{\link[=calcValue]{calcValue()}}.
}
}
\keyword{internal}
