% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo.R
\docType{package}
\name{rmo-package}
\alias{rmo-package}
\title{rmo: A package for the Marshall-Olkin distribution.}
\description{
The package contains fast implementations of sampling algorithms related to the Marshall-Olkin
distribution. It also contains an S4-class system for Bernstein functions which can be used to
create a large variety of Marshall-Olkin parameterisations.
}
\section{Bernstein functions}{

A \emph{Bernstein function} is a non-negative function with non-negative alternating iterative
differences. These functions can be used to generated parameterisations the extendible
Marshall-Olkin distributions.
\itemize{
\item An object that derives from \linkS4class{BernsteinFunction} can be used to generate the
Marshall-Olkin shock arrival intensities with \code{\link[=intensities]{intensities()}}. It can be used to generate
(scaled) exchangeable shock-size arrival intensities with \code{\link[=exIntensities]{exIntensities()}}.
\item Many families of Bernstein functions are available, e.g. \linkS4class{ParetoBernsteinFunction},
\linkS4class{ExponentialBernsteinFunction}, and \linkS4class{AlphaStableBernsteinFunction}.
\item Bernstein functions can be recombined by scaling, by summation or by composition, which can be
used to create new Bernstein functions with \linkS4class{ScaledBernsteinFunction} and
\linkS4class{SumOfBernsteinFunctions}.
}
}

\section{Sampling}{

Multiple sampling algorithms are provided. The semantic naming scheme \code{r*mo} is used, e.g.,
\itemize{
\item \code{\link[=rmo]{rmo()}} allows the simulation from Marshall-Olkin distributions. It has \emph{shock arrival
intensities} as input and uses the \emph{Arnold model} or \emph{exogenous shock model} for sampling; the
former can be used up until dimension \eqn{30}, but the latter should only be used in very
small dimensions.
\item \code{\link[=rexmo]{rexmo()}} allows the simulation from exchangeable Marshall-Olkin distributions. It has
\emph{exchangeable shock-size arrival intensities} as input and uses the \emph{Markovian death-set
model}, the \emph{Arnold model}, or the \emph{exogenous shock model} for sampling; for the latter two
models, the corresponding \emph{shock arrival intensities} are calculated and passed down to
\code{\link[=rmo]{rmo()}}.
\item \code{\link[=rextmo]{rextmo()}} allows the simulation from extendible Marshall-Olkin distributions. It has a
\emph{Bernstein function} as input and calculates the corresponding \emph{exchangeable shock-size arrival
intensities} and passes them down to \code{\link[=rexmo]{rexmo()}}.
\item \code{\link[=rpextmo]{rpextmo()}} allows the simulation from parametric families of extendible Marshall-Olkin
distributions. It has a \emph{killing-rate}, a \emph{drift}, a \emph{scaling factor}, a \emph{parameter vector},
and a \emph{family name} as input. All previous models can be used and the corresponding inputs are
calculated from the Bernstein function. For special families, which correspond to Lévy
exponents of compound Poisson subordinators, the \emph{Lévy frailty model} can be used for
simulations.
}
}

