% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo-package.R
\docType{package}
\name{rmo-package}
\alias{rmo-package}
\title{rmo: A package for simulating Marshall–Olkin distributions}
\description{
The rmo-package provides efficient sampling algorithms for the Marshall–Olkin
distribution and a flexible S4-class system for creating diverse
parametrizations.
}
\section{Sampling}{

Simulation algorithms are provided for various MO parametrizations. The
semantic naming scheme \code{r*mo} is used, e.g.,
\itemize{
\item \code{\link[=rpextmo]{rpextmo()}} allows to simulate from parametric families of extendible
Marshall–Olkin distributions. The function takes a \emph{killing-rate}, a
\emph{drift}, a \emph{scaling factor}, a \emph{parameter vector}, and a \emph{family name} as
input.
\item \code{\link[=rextmo]{rextmo()}} allows to simulate from extendible Marshall–Olkin
distributions. It takes a \emph{Bernstein function} as input.
\item \code{\link[=rexmo]{rexmo()}} allows to simulate from exchangeable Marshall–Olkin
distributions. It takes a vector of \emph{exchangeable shock-size arrival
intensities} as input.
\item \code{\link[=rmo]{rmo()}} allows to simulate from Marshall–Olkin distributions. It takes
vector of \emph{shock arrival intensities} as input and uses the \emph{Arnold model}
or \emph{exogenous shock model} for sampling; the former can be used up until
dimension \eqn{30}, but the latter should only be used in very small
dimensions.
}

The default simulation algorithm is the \emph{Markovian death-counting model}.
Dependent on the parametrization, other algorithms can be used, e.g., the
\emph{exogenous shock model}, the \emph{Arnold model}, or the \emph{Lévy-frailty model}.
}

\section{Bernstein functions}{

A \emph{Bernstein function} can be used to parametrize the \emph{extendible
Marshall–Olkin distribution}.
\itemize{
\item Many families of Bernstein functions are available,
e.g. \linkS4class{ParetoBernsteinFunction},
\linkS4class{ExponentialBernsteinFunction}, and
\linkS4class{AlphaStableBernsteinFunction}.
\item Bernstein functions can be recombined by scaling, by summation or by
composition, which can be used to create new Bernstein functions with
\linkS4class{ScaledBernsteinFunction}, \linkS4class{CompositeScaledBernsteinFunction}
and \linkS4class{SumOfBernsteinFunctions}.
\item An object that derives from \linkS4class{BernsteinFunction} can be used to
generate the Marshall–Olkin shock arrival intensities with \code{\link[=intensities]{intensities()}}.
It can be used to generate (scaled) exchangeable shock-size arrival
intensities with \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hsloot/rmo}
  \item Report bugs at \url{https://github.com/hsloot/rmo/issues}
}

}
\author{
\strong{Maintainer}: Henrik Sloot \email{henrik.sloot@gmail.com} (\href{https://orcid.org/0000-0001-8582-7776}{ORCID})

}
\keyword{internal}
