% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo}
\alias{rmo}
\title{Sample from a Marshall--Olkin distribution}
\usage{
rmo(n, d, intensities, method = c("AM", "ESM"))
}
\arguments{
\item{n}{an integer for the number of samples.}

\item{d}{an integer for the dimension of the sample.}

\item{intensities}{a numeric vector with the Marshall-Olkin shock arrival intensities.}

\item{method}{a character vector indicating which sampling algorithm should be used.
Use "AM" for the \emph{Arnold model} and "ESM" for the \emph{exogenous shock model}.}
}
\value{
\code{rmo} returns an \eqn{n \times d}{n x d} numeric matrix with the rows
corresponding to independent and identically distributed samples of a
\eqn{d} variate Marshall-Olkin distribution with parameters \code{intensities}.
}
\description{
Draws \code{n} independent samples from a \code{d}-variate Marshall-Olkin distribution
with shock arrival \code{intensities}.
}
\details{
\strong{Parametrisation}:
The Marshall--Olkin distribution has the survival function
\deqn{
  \bar{F}(t)
     = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 ,
}
for \emph{shock arrival intensities} \eqn{\lambda_I \geq 0},
\eqn{\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}}.

\strong{The shock arrival intensities}:
\itemize{
\item The shock arrival \code{intensities} must be stored in a vector of length
\eqn{2^d-1}.
\item A shock arrival intensity of zero corresponds to an almost surely infinite
shock arrival time.
\item We use a binary representation to map a non-empty subset \eqn{J}
of \eqn{\{ 1, \ldots, d\}}{{1, \ldots, d}} to integers \eqn{j} of
\eqn{1, \ldots, 2^d-1}.
In particular, \eqn{i} is a component in the set \eqn{J} corresponding to
the integer \eqn{j} if, and only if,
\eqn{j = \sum_{k=0}^\infty a_k 2^k}{\sum a[k] * 2^k}
and \eqn{a_{i-1} = 1}{a[i-1] = 1}.
}

The \strong{exogenous shock model} simulates a Marshall--Olkin distributed random
vector via exponentially distributed shock times for all non-empty subsets,
see \insertCite{@see pp. 104 psqq. @Mai2017a}{rmo} and
\insertCite{Marshall1967a}{rmo}.

The \strong{Arnold model} simulates a Marshall--Olkin distributed random variable
by simulating a marked homogeneous Poisson process and where the
inter-arrival times correspond to shock shock-arrival times and the marks to
the specific shocks, see \insertCite{@see Sec. 3.1.2 @Mai2017a}{rmo} and
\insertCite{Arnold1975a}{rmo}.
}
\examples{
rmo(10L, 2L, c(0.4, 0.3, 0.2))
rmo(10L, 2L, c(1, 1, 0))         ## independence
rmo(10L, 2L, c(0, 0, 1))         ## comonotone

rmo(10L, 2L, c(0.4, 0.3, 0.2), method = "ESM")
rmo(10L, 2L, c(1, 1, 0), method = "ESM")         ## independence
rmo(10L, 2L, c(0, 0, 1), method = "ESM")         ## comonotone

rmo(10L, 2L, c(0.4, 0.3, 0.2), method = "AM")
rmo(10L, 2L, c(1, 1, 0), method = "AM")         ## independence
rmo(10L, 2L, c(0, 0, 1), method = "AM")         ## comonotone

}
\references{
\insertAllCited{}
}
\seealso{
Other sampling-algorithms: 
\code{\link{rarmextmo_esm}()},
\code{\link{rexmo}()},
\code{\link{rextmo_lfm}()}
}
\concept{sampling-algorithms}
