% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo.R
\docType{package}
\name{rmo}
\alias{rmo}
\title{rmo: A package for the Marshall-Olkin distribution.}
\description{
The package contains an implementation of the sampling algorithms related
to the Marshall-Olkin distribution.
}
