% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo}
\alias{rmo}
\title{Sample from a Marshall–Olkin distribution}
\usage{
rmo(n, d, intensities, method = c("AM", "ESM"))
}
\arguments{
\item{n}{an integer for the \emph{number of samples}.}

\item{d}{an integer for the \emph{dimension}.}

\item{intensities}{a numeric vector for the \emph{shock arrival intensities}.}

\item{method}{a string indicating which sampling algorithm should be used.
Use "AM" for the \emph{Arnold model} and "ESM" for the \emph{exogenous shock model}. We recommend using
the \emph{ESM} only for small dimensions; the \emph{AM} can be used up until dimension \eqn{30}.}
}
\value{
\code{rmo} returns a numeric matrix with \eqn{n} rows and \eqn{d} columns with the rows
corresponding to iid distributed samples of a \eqn{d}-variate \emph{Marshall–Olkin distribution}
with \emph{shock arrival intensities} \code{intensities}.
}
\description{
Draws \eqn{n} iid samples from a \eqn{d}-variate \emph{Marshall–Olkin distribution}
parametrized by \emph{shock arrival intensities}.
}
\section{Parametrization}{
The \emph{Marshall–Olkin distribution} has the survival function
\deqn{
  \bar{F}(t)
     = \exp{\left\{ - \sum_{I} \lambda_I \max_{i \in I} t_i \right\}} , \quad t > 0 ,
}
for \emph{shock arrival intensities} \eqn{\lambda_I \geq 0},
\eqn{\emptyset \neq I \subseteq {\{ 1 , \ldots, d \}}}.
\subsection{Shock arrival intensities}{
\itemize{
\item The \emph{shock arrival intensities} must be stored in a vector of length \eqn{2^d-1}.
\item A \emph{shock arrival intensity} of zero corresponds to an almost surely infinite shock arrival
time.
\item We use a binary representation to map a non-empty subset \eqn{J} of
\eqn{\{ 1, \ldots, d\}}{{1, \ldots, d}} to integers \eqn{j} of \eqn{1, \ldots, 2^d-1}. In
particular, \eqn{i} is a component in the set \eqn{J}, corresponding to the integer \eqn{j},
if, and only if, \eqn{j = \sum_{k=0}^\infty a_k 2^k}{\sum a[k] * 2^k} and
\eqn{a_{i-1} = 1}{a[i-1] = 1}.
}
}
}

\section{Simulation algorithms}{
\subsection{Exogenous shock model}{

The \emph{exogenous shock model (ESM)} simulates a Marshall–Olkin distributed random vector via
exponentially distributed shock times for all non-empty subsets of components and defines each
component as the minimum of all shock times corresponding to a subset containing this component,
see \insertCite{@see pp. 104 psqq. @Mai2017a}{rmo} and \insertCite{Marshall1967a}{rmo}.
}

\subsection{Arnold model}{

The \emph{Arnold model (AM)} simulates a Marshall–Olkin distributed random vector by simulating a
marked homogeneous Poisson process where the inter-arrival times correspond to shock
inter-arrival times and the marks to the specific shocks; the process is stopped when all
components are \emph{hit} by a shock. As in the \emph{ESM}, each component is defined as the minimum of all
shock times corresponding to a subset containing this component, see
\insertCite{@see Sec. 3.1.2 @Mai2017a}{rmo} and \insertCite{Arnold1975a}{rmo}.
}
}

\examples{
rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4))
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0))         ## independence
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1))         ## comonotone

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "ESM")
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "ESM")         ## independence
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "ESM")         ## comonotone

rmo(10, 3, c(0.4, 0.4, 0.1, 0.4, 0.1, 0.1, 0.4), method = "AM")
rmo(10, 3, c(1, 1, 0, 1, 0, 0, 0), method = "AM")         ## independence
rmo(10, 3, c(0, 0, 0, 0, 0, 0, 1), method = "AM")         ## comonotone

}
\references{
\insertAllCited{}
}
\seealso{
Other sampling-algorithms: 
\code{\link{rexmo}()},
\code{\link{rextmo}()},
\code{\link{rpextmo}()}
}
\concept{sampling-algorithms}
