% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmo.R
\docType{package}
\name{rmo}
\alias{rmo}
\title{rmo: A package for the Marshall-Olkin distribution.}
\description{
The package contains an implementation of the sampling algorithms related to
the Marshall-Olkin distribution.
}
\section{Sampling}{

\itemize{
\item Sample with the \emph{exogenous shock model} from arbitrary Marshall-Olkin
distributions with \code{\link{rmo_esm}}. Not adviced for \code{d} much larger
than 2.
\item Sample with the \emph{Arnold model} from arbitrary Marshall-Olkin distributions
with \code{\link{rmo_arnold}}.
\item Sample with the \emph{exchangeable Markovian model} from exchangeable
Marshall-Olki distributions with \code{\link{rexmo_markovian}}.
\item Sample with the \emph{Lévy-frailty model for compound Poisson subordinators}
from the corresponding subclass of extendible Marshall-Olkin distributions
with \code{\link{rextmo_lfm}}.
\item Sample with the optimised version of the \emph{exogenous shock model} for the
Cuadras-Augé subclass with \code{\link{rcamo_esm}}.
}
}

