% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo_esm}
\alias{rmo_esm}
\title{Sampling from the exogenous shock model}
\usage{
rmo_esm(n, d, intensities)
}
\arguments{
\item{n}{number of samples}

\item{d}{dimension}

\item{intensities}{shock model intensity rates}
}
\value{
an \eqn{n \times d}{n x d} matrix with rows corresponding to the independent
samples of size \eqn{d}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate Marshall-Olkin distribution
with shock rates \code{intensities} using the \emph{exogenous shock model algorithm}.
The shock rates must be stored in a vector of length \eqn{2^d-1}. For more information
on this algorithm, see J.-F. Mai, M. Scherer, "Simulating Copulas", World Scientific (2017).
}
\examples{
# sample 10 times from a bivariate Marshall-Olkin distribution with parameters
# with c(1, 1, 1) using the exogenous shock model algorithm
rmo_esm(10, 2, c(1, 1, 1))

}
