% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sample.R
\name{rmo_esm_cuadras_auge}
\alias{rmo_esm_cuadras_auge}
\title{Sample from Cuadras-Auge distribution}
\usage{
rmo_esm_cuadras_auge(n, d, alpha, beta)
}
\arguments{
\item{n}{number of samples}

\item{d}{dimension}

\item{alpha}{rate of individual shocks}

\item{beta}{rate of global shock}
}
\value{
\code{rmo_esm_cuadras_auge} implements an optimized version of the
\emph{exogenous shock model} representation for the Cuadras-Augé family and
returns an \eqn{n \times d}{n x d} array matrix with rows corresponding to
the independent samples of size \eqn{d}.

\code{rmo_esm_cuadras_auge} implements an optimized version of the
\emph{exogenous shock model} representation for the Cuadras-Augé family and
returns an \eqn{n \times d}{n x d} array matrix with rows corresponding to
the independent samples of size \eqn{d}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate Cuadras-Augé distribution
with parameters \code{alpha} and \code{beta}.

Draws \code{n} independent samples from a \code{d} variate Cuadras-Augé distribution
with parameters \code{alpha} and \code{beta}.
}
\details{
\itemize{
\item \code{alpha} is the shock intensity of shocks that affect only single
components.
\item \code{beta} is the shock intensity of the global shock that affects all
components.
}

\itemize{
\item \code{alpha} is the shock intensity of shocks that affect only single
compoenents.
\item \code{beta} is the shock intensity of the global shock that affects all
components.
}
}
\examples{
rmo_esm_cuadras_auge(10L, 2L, 0.5, 0.2)
rmo_esm_cuadras_auge(10L, 2L, 0, 1)      ## comonotone
rmo_esm_cuadras_auge(10L, 2L, 1, 0)      ## independence

rmo_esm_cuadras_auge(10L, 2L, 0.5, 0.2)
rmo_esm_cuadras_auge(10L, 2L, 0, 1)      ## comonotone
rmo_esm_cuadras_auge(10L, 2L, 1, 0)      ## independence

}
\seealso{
\code{\link{rmo_esm}}

\code{\link{rmo_esm}}

Other samplers: 
\code{\link{rmo_esm}()},
\code{\link{rmo_ex_arnold}()},
\code{\link{rmo_lfm_cpp}()}

Other samplers: 
\code{\link{rmo_esm}()},
\code{\link{rmo_ex_arnold}()},
\code{\link{rmo_lfm_cpp}()}
}
\concept{samplers}
