% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo_esm_cuadras_auge}
\alias{rmo_esm_cuadras_auge}
\title{Sampler for Cuadras-Augé distribution}
\usage{
rmo_esm_cuadras_auge(n, d, alpha, beta)
}
\arguments{
\item{n}{number of samples}

\item{d}{dimension}

\item{alpha}{rate of individual shocks}

\item{beta}{rate of global shock}
}
\value{
An \eqn{n \times d}{n x d} array matrix with rows corresponding to
the independent samples of size \eqn{d}.
}
\description{
Draws \code{n} independent samples from a \code{d} variate Cuadras-Augé distribution
with the \emph{exogenous shock model}. This is a specialised version of
\code{rmo_esm}.
}
\examples{
rmo_esm_cuadras_auge(10L, 2L, 0.5, 0.2)
rmo_esm_cuadras_auge(10L, 2L, 0, 1)      ## comonotone
rmo_esm_cuadras_auge(10L, 2L, 1, 0)      ## independence

}
