% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rmo_lfm_cpp}
\alias{rmo_lfm_cpp}
\title{Sampler for LFM with CPP subordinator}
\usage{
rmo_lfm_cpp(n, d, rate, rate_killing, rate_drift, rjump_name, rjump_arg_list)
}
\arguments{
\item{n}{number of samples}

\item{d}{dimension}

\item{rate}{rate of CPP subordinator}

\item{rate_killing}{killing rate of CPP subordinator}

\item{rate_drift}{drift rate of CPP subordinator}

\item{rjump_name}{name of jump sampling function}

\item{rjump_arg_list}{list with named arguments for jump sampling function}
}
\value{
An \verb{n x d} array with \code{n} independent samples of the corresponding
\code{d} variate extendible Marshall-Olkin distribution.
}
\description{
Sampler for LFM with CPP subordinator
}
\examples{
rmo_lfm_cpp(10L, 2L, 0.5, 0, 0, "rposval", list("value"=1))
rmo_lfm_cpp(10L, 2L, 0.5, 0, 0, "rexp", list("rate"=2))

}
