% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ScaledBernsteinFunction.R
\docType{class}
\name{ScaledBernsteinFunction-class}
\alias{ScaledBernsteinFunction-class}
\alias{ScaledBernsteinFunction}
\title{Class for scaled Bernstein functions}
\description{
Bernstein functions are stable under (nonegative) scalar multiplication,
i.e. if \eqn{\psi} is a Bernstein function and \eqn{\lambda \geq 0} , then
\deqn{
  x \mapsto \lambda \psi(x), x > 0,
}
is also a Bernstein function.
}
\section{Slots}{

\describe{
\item{\code{scale}}{The scalar factor with which the original Bernstein function
is to be multiplied.}

\item{\code{original}}{The original Bernstein function which is to be multiplied.}
}}

\examples{
ScaledBernsteinFunction()
original_bf <- AlphaStableBernsteinFunction(alpha = 0.5)
ScaledBernsteinFunction(scale = 2, original = original_bf)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{SumOfBernsteinFunctions-class}}

Other Bernstein function transformer classes: 
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{SumOfBernsteinFunctions-class}}
}
\concept{Bernstein function classes}
\concept{Bernstein function transformer classes}
