% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions}
\title{Class for the \emph{sum of two Bernstein functions}}
\description{
Bernstein functions are stable under addition, i.e. if \eqn{\psi_1} and
\eqn{\psi_2} are two Bernstein functions, then
\deqn{
  x \mapsto \psi_1(x) + \psi_2(x) , x>0,
}
is also a Bernstein function.
}
\section{Slots}{

\describe{
\item{\code{first}}{The first summand (Bernstein function).}

\item{\code{second}}{The second summand (Bernstein function).}
}}

\examples{
SumOfBernsteinFunctions()
first_bf <- LinearBernsteinFunction(scale=0.2)
second_bf <- AlphaStableBernsteinFunction(alpha=0.5)
SumOfBernsteinFunctions(first=first_bf, second=second_bf)

}
\seealso{
\linkS4class{BernsteinFunction}
}
