% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-SumOfBernsteinFunctions.R
\docType{class}
\name{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions}
\title{Class for sums of two Bernstein functions}
\description{
Bernstein functions are stable under addition, i.e. if \eqn{\psi_1} and
\eqn{\psi_2} are two Bernstein functions, then
\deqn{
  x \mapsto \psi_1(x) + \psi_2(x) , x>0,
}
is also a Bernstein function.
}
\section{Slots}{

\describe{
\item{\code{first}}{The first summand (derived from \linkS4class{BernsteinFunction}).}

\item{\code{second}}{The second summand (derived from \linkS4class{BernsteinFunction}).}
}}

\examples{
# Create an object of class SumOfBernsteinFunctions
SumOfBernsteinFunctions()
SumOfBernsteinFunctions(
  first = LinearBernsteinFunction(scale = 0.2),
  second = AlphaStableBernsteinFunction(alpha = 0.5)
)
}
\seealso{
\code{\link[=valueOf]{valueOf()}}, \code{\link[=calcShockArrivalIntensities]{calcShockArrivalIntensities()}},
\code{\link[=calcExShockArrivalIntensities]{calcExShockArrivalIntensities()}}, \code{\link[=calcExShockSizeArrivalIntensities]{calcExShockSizeArrivalIntensities()}},
\code{\link[=calcMDCMGeneratorMatrix]{calcMDCMGeneratorMatrix()}}, \code{\link[=rextmo]{rextmo()}}, \code{\link[=rpextmo]{rpextmo()}}

Other Bernstein function classes: 
\code{\link{AlphaStableBernsteinFunction-class}},
\code{\link{BernsteinFunction-class}},
\code{\link{CompleteBernsteinFunction-class}},
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConstantBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ExponentialBernsteinFunction-class}},
\code{\link{GammaBernsteinFunction-class}},
\code{\link{InverseGaussianBernsteinFunction-class}},
\code{\link{LevyBernsteinFunction-class}},
\code{\link{LinearBernsteinFunction-class}},
\code{\link{ParetoBernsteinFunction-class}},
\code{\link{PoissonBernsteinFunction-class}},
\code{\link{ScaledBernsteinFunction-class}}

Other Bernstein function transformer classes: 
\code{\link{CompositeScaledBernsteinFunction-class}},
\code{\link{ConvexCombinationOfBernsteinFunctions-class}},
\code{\link{ScaledBernsteinFunction-class}}
}
\concept{Bernstein function classes}
\concept{Bernstein function transformer classes}
