% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/allShow-S4.R, R/valueOf-S4.R
\docType{class}
\name{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions-class}
\alias{SumOfBernsteinFunctions}
\alias{initialize,SumOfBernsteinFunctions-method}
\alias{initialize,SumOfBernsteinFunctions,ANY-method}
\alias{show,SumOfBernsteinFunctions-method}
\alias{valueOf,SumOfBernsteinFunctions-method}
\alias{valueOf,SumOfBernsteinFunctions,ANY-method}
\title{Class for \emph{input scaled Bernstein functions}}
\usage{
\S4method{initialize}{SumOfBernsteinFunctions}(.Object, first, second)

\S4method{show}{SumOfBernsteinFunctions}(object)

\S4method{valueOf}{SumOfBernsteinFunctions}(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{first}{Derives from \linkS4class{BernsteinFunction}.}

\item{second}{Derives from \linkS4class{BernsteinFunction}.}

\item{object}{An object deriving from \linkS4class{LevyBernsteinFunction}
(for \code{levyDensity}) or \linkS4class{CompleteBernsteinFunction}
(for \code{stieltjesDensity}).}

\item{x}{Non-negativ numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{j} in
the representation).}

\item{n, k}{Non-negative numbers for the binomial factors.}

\item{cscale}{Positive number, the composit scaling factor.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}
}
\description{
Class for the \emph{sum of two Bernstein functions}
}
\details{
Bernstein functions are stable under addition, i.e. if \eqn{\psi_1} and
\eqn{\psi_2} are two Bernstein functions, then
\deqn{
  x \mapsto \psi_1(x) + \psi_2(x) , x>0,
}
is also a Bernstein function.
}
\section{Functions}{
\itemize{
\item \code{initialize,SumOfBernsteinFunctions-method}: Constructor

\item \code{show,SumOfBernsteinFunctions-method}: Display the object.

\item \code{valueOf,SumOfBernsteinFunctions-method}: Calculates the iterated differences of the Bernstein function, see \code{\link[=valueOf]{valueOf()}}
}}

\section{Slots}{

\describe{
\item{\code{first}}{The first summand (derived from \linkS4class{BernsteinFunction}).}

\item{\code{second}}{The second summand (derived from \linkS4class{BernsteinFunction}).}
}}

\examples{
SumOfBernsteinFunctions()
first_bf <- LinearBernsteinFunction(scale = 0.2)
second_bf <- AlphaStableBernsteinFunction(alpha = 0.5)
SumOfBernsteinFunctions(first = first_bf, second = second_bf)
}
\seealso{
\linkS4class{BernsteinFunction},
\code{\link[=valueOf]{valueOf()}}
}
