% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\docType{methods}
\name{valueOf}
\alias{valueOf}
\alias{valueOf,LinearBernsteinFunction,numeric,integer-method}
\alias{valueOf,LinearBernsteinFunction,numeric,integer,ANY-method}
\alias{valueOf,ConstantBernsteinFunction,numeric,integer-method}
\alias{valueOf,ConstantBernsteinFunction,numeric,integer,ANY-method}
\alias{valueOf,ScaledBernsteinFunction,numeric,integer-method}
\alias{valueOf,ScaledBernsteinFunction,numeric,integer,ANY-method}
\alias{valueOf,SumOfBernsteinFunctions,numeric,integer-method}
\alias{valueOf,SumOfBernsteinFunctions,numeric,integer,ANY-method}
\alias{valueOf,PoissonBernsteinFunction,numeric,integer-method}
\alias{valueOf,PoissonBernsteinFunction,numeric,integer,ANY-method}
\alias{valueOf,AlphaStableBernsteinFunction,numeric,integer-method}
\alias{valueOf,AlphaStableBernsteinFunction,numeric,integer,ANY-method}
\title{Returns values for Bernstein functions}
\usage{
valueOf(object, x, difference_order)

\S4method{valueOf}{LinearBernsteinFunction,numeric,integer}(object, x, difference_order = 0)

\S4method{valueOf}{ConstantBernsteinFunction,numeric,integer}(object, x, difference_order = 0)

\S4method{valueOf}{ScaledBernsteinFunction,numeric,integer}(object, x, difference_order = 0)

\S4method{valueOf}{SumOfBernsteinFunctions,numeric,integer}(object, x, difference_order = 0)

\S4method{valueOf}{PoissonBernsteinFunction,numeric,integer}(object, x, difference_order = 0)

\S4method{valueOf}{AlphaStableBernsteinFunction,numeric,integer}(object, x, difference_order = 0)
}
\arguments{
\item{object}{The Bernstein function object.}

\item{x}{The value at which the Bernstein function is to be evaluated.}

\item{difference_order}{The order of the alternating iterated forward
differences taken on the Bernstein function (\eqn{k} in
the representation).}
}
\description{
This method allows you to calculate the values for a Bernstein function
and its higher-order, alternating iterated forward differences, i.e.
\deqn{
  {(-1)}^{k} \Delta^k \psi(x), x > 0.
}
}
\seealso{
\linkS4class{BernsteinFunction}

\linkS4class{LinearBernsteinFunction}

\linkS4class{ConstantBernsteinFunction}

\linkS4class{ScaledBernsteinFunction}

\linkS4class{SumOfBernsteinFunctions}

\linkS4class{PoissonBernsteinFunction}

\linkS4class{AlphaStableBernsteinFunction}
}
