% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-valueOf.R
\name{valueOf}
\alias{valueOf}
\title{Evaluate Bernstein Function differences}
\usage{
valueOf(object, x, difference_order = 0L, n = 1L, k = 0L, cscale = 1, ...)
}
\arguments{
\item{object}{An object deriving from the class \linkS4class{BernsteinFunction}.}

\item{x}{A nonnegative numeric vector at which the iterated difference of
the Bernstein function is evaluated.}

\item{difference_order}{A nonnegative integer with the order of the
alternating iterated forward differences taken on the Bernstein function.}

\item{n, k}{Nonnegative numbers for the binomial factor.}

\item{cscale}{A positive number for the \emph{composite scaling factor}.}

\item{...}{Pass-through parameter.}
}
\description{
Calculate the values for a Bernstein function and its higher-order,
alternating iterated forward differences, possibly scaled by a binomial
coefficient, i.e.,
\deqn{
  \binom{n}{k} {(-1)}^{j-1} \Delta^{j}{ \psi(c x) } ,
    \quad x > 0 .
}
The evaluation of Bernstein functions using this formula is usually not
numerically stable. Consequently, the various alternative approaches are used
dependent on the class of the Bernstein function.
}
\examples{
bf <- AlphaStableBernsteinFunction(alpha = 0.7)
valueOf(bf, 1:5)
}
\seealso{
Other Bernstein function generics: 
\code{\link{calcExShockArrivalIntensities}()},
\code{\link{calcExShockSizeArrivalIntensities}()},
\code{\link{calcMDCMGeneratorMatrix}()},
\code{\link{calcShockArrivalIntensities}()},
\code{\link{getDefaultMethodString}()},
\code{\link{getLevyDensity}()},
\code{\link{stieltjesDensity}()},
\code{\link{valueOf0}()}
}
\concept{Bernstein function generics}
